### 2019-07-15 #1571 @whedon
[PRE REVIEW]: cspy: A Python package with a collection of algorithms for the (Resource) Constrained Shortest Path problem
**Submitting author:** @torressa (<a href="http://orcid.org/0000-0002-2894-9432">David Torres Sanchez</a>)
**Repository:** <a href="https://github.com/torressa/cspy" target ="_blank">https://github.com/torressa/cspy</a>
**Version:** v0.0.8
