+++ paper/paper.bib
@@ -9,7 +9,7 @@
 
 @article{righini2006,
   title={Symmetry helps: Bounded bi-directional dynamic programming for the elementary shortest path problem with resource constraints},
-  doi = {10.1016/j.disopt.2006.05.007}
+  doi = {10.1016/j.disopt.2006.05.007},
   author={Righini, Giovanni and Salani, Matteo},
   journal={Discrete Optimization},
   volume={3},
