### 2019-07-15 #1571 @danielskatz
👋 @torressa - I think https://github.com/torressa/cspy/pull/12 will fix the build error.  Please merge it and then in a new comment here, enter ```@whedon generate pdf``` Please iterate until the paper builds :)
