### 2019-07-16 f73d73d
removed unnecessary inputs; paper cleanup
