+++ paper/paper.bib
@@ -196,7 +196,7 @@ year = {2019}
   author = {Weyens, Toon},
   title = {{\texttt{pylgrim}-- Elementary Shortest Path Problem with or without Resource Constraint}},
   year =  {2018},
-  url = {\url{https://github.com/ToonWeyens/pylgrim}}
+  url = {https://github.com/ToonWeyens/pylgrim}
 }
 @article{Feillet2004,
 author = {Feillet, Dominique and Dejax, Pierre and Gendreau, Michel and Gueguen, Cyrille},
