### 2019-07-18 1a56da7
updated README; added CHANGELOG.md; codeacy fixes
