+++ CHANGELOG.md
@@ -1,0 +1,50 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [Unreleased]
+
+## [0.0.8] - 15/07/2019
+### Added
+
+- Generic resource extension functions options.
+
+### Changed
+
+- numpy.array integration.
+
+## [0.0.5] - 9/07/2019
+### Added
+
+- ``PSOLGENT``.
+- ``GreedyElim`` simple test.
+
+### Changed
+
+- Fixed prune_graph preprocessing routine.
+- YAPF google style.
+
+## [0.0.3] - 9/07/2019
+### Added
+
+- ``GRASP``.
+
+### Changed
+
+- Documentation updates.
+- Updated README.
+- Removed duplicate code in `tabu.py` and `greedy_elimination.py`.
+
+## [0.0.1] - 1/07/2019
+### Added
+
+- assertLogs tests for bidirectional algorithm classification.
+- Personal MIT LICENSE.
+- ``GreedyElim`` Procedure.
+
+### Changed
+
+- Documentation updates.
+- Docstring modifications to include maths.
+- Updated README.
