### 2019-07-18 497efde
added comparison links in CHANGELOG
