+++ CHANGELOG.md
@@ -8,19 +8,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.0.8] - 15/07/2019
 ### Added
 
+
 - Generic resource extension functions options.
 
 ### Changed
+
 
 - numpy.array integration.
 
 ## [0.0.5] - 9/07/2019
 ### Added
 
+
 - ``PSOLGENT``.
 - ``GreedyElim`` simple test.
 
 ### Changed
+
 
 - Fixed prune_graph preprocessing routine.
 - YAPF google style.
@@ -28,9 +32,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.0.3] - 9/07/2019
 ### Added
 
+
 - ``GRASP``.
 
 ### Changed
+
 
 - Documentation updates.
 - Updated README.
@@ -39,12 +45,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.0.1] - 1/07/2019
 ### Added
 
+
 - assertLogs tests for bidirectional algorithm classification.
 - Personal MIT LICENSE.
 - ``GreedyElim`` Procedure.
 
 ### Changed
 
+
 - Documentation updates.
 - Docstring modifications to include maths.
 - Updated README.
+
+
+[Unreleased]: https://github.com/torressa/cspy/compare/v0.0.8...HEAD
+[0.0.8]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.7...v0.0.8
+[0.0.5]: https://github.com/torressa/cspy/compare/0.0.3...0.0.5
+[0.0.3]: https://github.com/torressa/cspy/compare/0.0.1...0.0.3
+[0.0.1]: https://github.com/torressa/cspy/releases/tag/v0.0.1
+
