### 2019-07-18 a1c49fd
added comparison links in CHANGELOG
