### 2019-07-18 b9ba145
added Codacy badge
