### 2019-07-18 a815ca8
updated PSOLGENT default inputs and test
