### 2019-07-22 #1571 @poulson
Hi @danielskatz -- I am able to serve as an editor and apologize for the delay.
