### 2019-07-27 007038f
updated README (joss badge + test instructions); updated docstrings
