### 2019-07-27 d4c68fb
updated README (joss badge + test instructions); updated docstrings
