### 2019-07-27 236fee8
updated readme and requirements with dev modules
