+++ requirements.txt
@@ -1,2 +1,3 @@
 networkx>=2.2
 numpy>=1.13.3
+coverage>=4.5.3
