### 2019-07-27 9ae42b2
updated psolgent (added seed to avoid test errors)
