+++ .circleci/config.yml
@@ -2,9 +2,9 @@
 #
 # Check https://circleci.com/docs/2.0/language-python/ for more details
 #
-version: 2.1
+version: 2
 orbs:
-  codecov: codecov/codecov@1.0.5
+  codecov: codecov/codecov@1.0.0
 jobs:
   build:
     docker:
