### 2019-07-27 73fda9c
circleci fix attempt, manually install codecov in workflow
