### 2019-07-30 #1571 @poulson
Hi Camille (@camillescott) -- Given your expertise in C++ and Cython, as well as graphs, complex networks, and workflows, I think you would be an ideal reviewer for this submission. Would you have the time and interest in doing so?

Hi @gabrielspadon -- You listed very similar expertise as Valentino, and I believe you would similarly be a perfect complement to the C++ expertise of Rodrigo in this review. Do you have the time and interest to help with the review?
