### 2019-07-31 a6acd4a
updated readme with issue templates
