### 2019-08-12 #1571 @poulson
Hi Scott (@stsievert): I apologize if I've asked you to review too many times, but you were recommended as a relevant reviewer for this paper. Would you have the time and interest in doing so?

Hi Oscar (@oesteban): David recommended you as a reviewer for this paper; would you be interested and confident in serving as a reviewer?

Hi Sarat (@sarats): You were recommended as an expert reviewer for this paper on Constrained Shortest Path problems: do you have the time and interest to serve as a reviewer?

This corresponding repository is available here:
https://github.com/torressa/cspy
