### 2019-08-14 ecc3e07
updated docs
