+++ docs/how_to.rst
@@ -66,7 +66,7 @@ For assumption 2, if resource extension functions are additive, these are easily
 REFs
 ~~~~
 
-Additive resource extension functions (REFs), are implemented by default in all the algorithms. However, you can use your own custom REFs. For theoretical information on what REFs we refer you to the paper by `Inrich 2005`_.
+Additive resource extension functions (REFs), are implemented by default in all the algorithms. However, you can use your own custom REFs. For theoretical information on what REFs are we refer you to the paper by `Inrich 2005`_.
 
 Practically, a custom REF will need two inputs: ``res``, a cumulative resource array, and ``edge``, an edge to consider for the extension of the current partial path. This function will be called every time the algorithms wish to consider and edge as part of the shortest path.
 
@@ -78,14 +78,14 @@ As an example, the following function would be valid:
 
         def REF_CUSTOM(cumulative_res, edge):
         	new_res = array(cumulative_res)
-        	# your filtering criteria that changes the elements of arr
+        	# your filtering criteria that changes the elements of new_res
         	# For example:
         	head_node, tail_node = edge[0:2]
         	if head_node != tail_node:
         		new_res[0] = 0
         	else:
         		new_res[0] = new_res[0] + 1
-        	return arr
+        	return new_res
 
 Your custom REF can then be passed with this format, into the algorithm of choice using the ``REF`` argument (see individual algorithms for details). Note that for the :class:`BiDirectional` algorithm, due to the properties of the algorithm, if you want to use this feature, you have to pass two custom REFs: one for the forward search and one for the backward search. Where the backward REF has to be the inverse of the forward REF, otherwise the algorithm will not return a meaningful path (`Tilk et al 2017`_).
 
