### 2019-08-14 629dff9
updated readme (spelling mistakes)
