### 2019-08-15 #1571 @stsievert
I can review this. I have moderate interest, but no experience in shortest path algorithms. I can reserve some time for this.
