### 2019-08-16 #1571 @poulson
Thank you Oscar (@oesteban) and Scott (@stsievert). I will go ahead and add you two and start the review. Moderate interest is...an interesting way to put it...but I don't want to delay the review any further.
