### 2019-08-16 #1655 @whedon
[REVIEW]: cspy: A Python package with a collection of algorithms for the (Resource) Constrained Shortest Path problem
**Submitting author:** @torressa (<a href="http://orcid.org/0000-0002-2894-9432">David Torres Sanchez</a>)
**Repository:** <a href="https://github.com/torressa/cspy" target ="_blank">https://github.com/torressa/cspy</a>
**Version:** v0.1.0

    * see https://github.com/torressa/cspy/issues/14
    * see https://github.com/torressa/cspy/issues/15
    * (in-depth example provided on GitHub; as of 2020-01-08 simple example on documentation not provided)
    * see https://github.com/torressa/cspy/issues/16
