### 2019-08-20 cb3d568
added random_state (thanks @stsievert, as per https://github.com/torressa/cspy/commit/9ae42b24341b870a75d8c14c503f28d600c616bc#commitcomment-34745792)
