### 2019-08-20 cedcf46
removed unnecessary import, updated ci version
