+++ .circleci/config.yml
@@ -2,7 +2,7 @@
 #
 # Check https://circleci.com/docs/2.0/language-python/ for more details
 #
-version: 2
+version: 2.1
 orbs:
   codecov: codecov/codecov@1.0.0
 jobs:
