### 2019-08-20 #1655 @stsievert
Thanks for the submission @torressa! This package looks pretty good.

The biggest improvements I see are documentation, specifically with the statement of need and a use of the software.

* I think the "statement of need" could use work, in the paper and especially in the documentation. I don't see any motivating reasons on why I, a grad student studying optimization (but not an expert on shortest path problems) should pick up cspy.
    * Possibly relevant questions: When is the constrained shortest path problem relevant? What concrete problems can it be applied to? What features does cspy provide, and why is it unique?
    * The provided statement of need is a lot clearer in paper than the documentation.
* I'd also like to more in-depth use of this software. Right now, I can only see some toy examples in the docstrings. Ideally, I think this would be best with an example with real data.
    * One question I have: how are constraints from real problems coded in cspy? I know cspy can be used in the vehicle routing problem with time windows (desaulniers2008tabu), but don't know how to code the time constraints in cspy.

I can point to some examples of other projects with good motivation/statement of needs if you'd like.

I feel these points would increase my inclination to use or recommend this software (though I am not a member of the target audience, the operational research community).

Some other comments:

* Could you add a point on how to contribute code? There are some good instructions in README.md for issues, but it's a little slim on "contributing to the software" and "seeking support".
* @poulson should books have DOIs? They list the book "Computers and Intractability; A Guide to the Theory of NP-Completeness" as a reference without a DOI.

I might have some more comments in the future.
