### 2019-08-21 #1655 @torressa
Thanks @stsievert for the comments and your prompt review :)
Really appreciate the comments. I'll raise them as separate issues and work on them in the coming days.
