### 2019-08-21 torressa/cspy#14 @torressa
JOSS REVIEW. DOCUMENTATION: Statement of need
From https://github.com/openjournals/joss-reviews/issues/1655#issuecomment-523107985

- [x] I think the "statement of need" could use work, in the paper and especially in the documentation. I don't see any motivating reasons on why I, a grad student studying optimization (but not an expert on shortest path problems) should pick up cspy.

- Possibly relevant questions: When is the constrained shortest path problem relevant? What concrete problems can it be applied to? What features does cspy provide, and why is it unique?
- The provided statement of need is a lot clearer in paper than the documentation.
