### 2019-08-21 torressa/cspy#15 @torressa
JOSS REVIEW. DOCUMENTATION: In-depth Use
From https://github.com/openjournals/joss-reviews/issues/1655#issuecomment-523107985

- [x] I'd also like to more in-depth use of this software. Right now, I can only see some toy examples in the docstrings. Ideally, I think this would be best with an example with real data.

One question I have: how are constraints from real problems coded in cspy? I know cspy can be used in the vehicle routing problem with time windows (desaulniers2008tabu), but don't know how to code the time constraints in cspy.

Answer: *Will be clear with a working example*
