### 2019-08-21 torressa/cspy#16 @torressa
JOSS REVIEW. DOCUMENTATION: Contributing and Seeking Support
From https://github.com/openjournals/joss-reviews/issues/1655#issuecomment-523107985

- [x] Could you add a point on how to contribute code? There are some good instructions in README.md for issues, but it's a little slim on "contributing to the software" and "seeking support".
