### 2019-08-21 723ce86
updated readme, removed hard-coded seed (https://github.com/torressa/cspy/commit/cb3d568ac0306065ef035044f5c4bcc55ed70ee8#commitcomment-34762935)
