### 2019-08-21 #1655 @stsievert
Thanks for opening those issues @torressa – that's useful. They provide a channel for more detailed feedback and any questions you have.
