### 2019-08-21 torressa/cspy#14 @stsievert
Thanks for opening this issue @torressa. I left the comment in https://github.com/torressa/cspy/issues/14#issue-483283722 (which I think is too strongly worded).

I think the statement of need in the paper explains the software to the target audience (though it's a very technical description). I think explaining some of the technical terms and/or providing a short example would make it more accessible and useful to a broader audience (e.g., an operational researcher with a different technical focus).

> cspy is an open source Python package that gathers some algorithms to solve the (resource) Constrained Shortest Path problem.
>
> By setting different options when calling the algorithms, one can have up to five different algorithms:
>
> [list of algorithms]
>
> Features implemented include:  generic resource extension functions (Inrich 2005) (not restricted to additive resources), generic resource consumptions (not restricted to non-negative values), and, increased efficiency (when compared to other implementations of monodirectional algorithms).
>
> —cspy documentation

I don't think the documentation explains the software to the target audience (the operational research community) because that audience includes people who are not CSP experts. It requires that the person reading the documentation to know what the CSP problem is (and when it's relevant) and specific details on the features.
