### 2019-08-22 torressa/cspy#14 @torressa
That totally makes sense now, thanks. I'll work on making a working example in the repo and a simple example and
