### 2019-09-03 #1655 @oesteban
@poulson Sorry, I thought I wasn't assigned this one (https://github.com/openjournals/joss-reviews/issues/1571#issuecomment-520653198). I'm still pretty swamped so I can't take this for review - I will for sure hold it up for long.
