### 2019-09-22 #1655 @stsievert
@torressa I've updated https://github.com/openjournals/joss-reviews/issues/1655#issuecomment-523107985 to include a note of the API documentation.
