### 2019-10-30 #1655 @danielskatz
👋 I'm just checking in on this as the associate-editor-in-chief on duty this week ... What's the status here?
