### 2019-11-03 torressa/cspy#15 @stsievert
Here's the information I see in different examples:

* https://cspy.readthedocs.io/en/latest/how_to.html
    * How to create a graph w/ Networkx
    * How to create an additive resource extension functions (REFs) (but not how to use it, or when it's useful)
* Docstrings (e.g., https://cspy.readthedocs.io/en/latest/api/cspy.PSOLGENT.html)
    * How to get a path from an algorithm

I'd like to see an in-depth use, preferably with real data. With the graph, how would I specify the weights, REFs and `max_res`/`min_res`? What are the tradeoffs between algorithms?

I think I understand resource usage (and the variables `min_res`/`max_res`/`n_res`/`res_cost`). Without reading the papers, I don't understand how resource usage relates to graph weights.
