### 2019-11-03 #1655 @stsievert
:+1: Thanks for checking in @danielskatz! I've been waiting until some points in https://github.com/openjournals/joss-reviews/issues/1655#issuecomment-523107985 and issues I've filed to be resolved.

I have filed some issues on the source repository (https://github.com/torressa/cspy/issues/14, https://github.com/torressa/cspy/issues/15, https://github.com/torressa/cspy/issues/16). I'm don't think this paper can be accepted until some of these issues are resolved.
