### 2019-11-03 82d17ff
Added example in the paper (issue #14)
