+++ paper/paper.bib
@@ -1,3 +1,10 @@
+@Misc{jpath,
+  author = {Torres Sanchez, David},
+  title = {{jpath: cspy example with Jane the postwoman}},
+  year =  {2019},
+  url = {\url{https://github.com/torressa/jpath}}
+}
+
 @book{gary1979,
  author = {Garey, Michael R. and Johnson, David S.},
  title = {Computers and Intractability; A Guide to the Theory of NP-Completeness},
