+++ paper/paper.md
@@ -18,13 +18,13 @@ bibliography: paper.bib
 
 # Introduction
 
-When solving the shortest path problem and considering multiple operational restrictions, one may resort to the (resource) constrained shortest path (CSP) problem. 
-In the applied column generation framework, particularly in the scheduling related literature, the CSP problem is commonly employed to generate columns. It consists, as its name suggests, in finding, among all paths, the shortest path from source to sink nodes that satisfies a set of constraints for a defined set of resources. 
+When solving the shortest path problem and considering multiple operational restrictions, one may resort to the (resource) constrained shortest path (CSP) problem.
+It consists, as its name suggests, in finding, among all paths, the shortest path from source to sink nodes that satisfies a set of constraints for a defined set of resources.
+Such set of resources and the way they evolve throughout the path are user defined and controlled. This allows the modelling of a wide variety of problems including: the vehicle routing problem with time windows, the technician routing and scheduling problem, the capacitated arc-routing problem, on-demand transportation systems, and, airport ground movement [@Desrochers1988; @Feillet2004; @Inrich2006; @Righini2008; @Bode2014; @Chen2016; @GARAIX201062; @Tilk2017; @Zamorano2017].
 
 ``cspy`` is a Python package that allows you to solve instances of the CSP problem using up to eight different algorithms.
 
-The CSP problem was first introduced as a subproblem for the bus driver scheduling problem.
-These have been shown to be applicable to a wide variety of problems including: the vehicle routing problem with time windows, the technician routing and scheduling problem, the capacitated arc-routing problem, on-demand transportation systems, and, airport ground movement [@Desrochers1988; @Feillet2004; @Inrich2006; @Righini2008; @Bode2014; @Chen2016; @GARAIX201062; @Tilk2017; @Zamorano2017]. 
+``cspy`` is of interest to the operational research community and others that wish to solve an instance of the CSP problem.
 
 # Algorithms
 
@@ -46,11 +46,124 @@ Some metaheuristics developed include, Tabu search [@desaulniers2008tabu], hybri
 - Tabu search. Adapted from @desaulniers2008tabu;
 - Greedy elimination procedure;
 - Greedy Randomised Adaptive Search Procedure (GRASP). Adapted from @Ferone2019;
-- Particle Swarm Optimization with combined Local and Global Expanding Neighborhood Topology (PSOLGENT) [@marinakis2017hybrid].
+- Particle Swarm Optimization with combined Local and Global Expanding Neighborhood Topology (PSOLGENT) @marinakis2017hybrid.
 
-The implementations allow for custom resource extension functions (REFs) to be used. Hence, allowing for generic consumption of the resources through the graph, not restricted to additive and subtractive REFs [@irnich2008]. 
+# Features
 
-``cspy`` is of interest to the operational research community and others that wish to solve an instance of the CSP problem.
+A key component of a CSP problem is to define a set of resources and a function to apply for these to evolve through the graph. Such functions are typically referred to as resource extension functions (REFs). The simplest REF is an additive one, where every time an edge is traversed a constant unit of a certain resource is consumed. However, custom and more generic REFs can be used.
+
+``cspy`` allows for custom and generic REFs to be used. Hence, allowing for a custom consumption of the resources through the graph.
+
+# Example
+
+Consider the following problem. 
+
[…]
