### 2019-11-03 #1655 @torressa
Sorry, was on holiday then swamped with new job. Will get going over the next weekends.
Thanks @stsievert  for your feedback!
