### 2019-11-03 aee71bf
Removed \url in last bib entry
