+++ paper/paper.bib
@@ -2,7 +2,7 @@
   author = {Torres Sanchez, David},
   title = {{jpath: cspy example with Jane the postwoman}},
   year =  {2019},
-  url = {\url{https://github.com/torressa/jpath}}
+  url = {https://github.com/torressa/jpath}
 }
 
 @book{gary1979,
