### 2019-11-12 #1655 @poulson
Thank you, @stsievert -- it looks like the rest of the review depends upon you!

@torressa -- am I understanding that you are expecting to need one more weekend to respond to @stsievert's issues?
