### 2019-11-13 #1655 @torressa
@poulson That's right!
I made some changes in the paper but docs and actual functioning example need to be updated.
