### 2019-11-14 b372429
JOSS Review. Issue #16
