### 2019-11-14 d47e57e
updated readme, enumurate imports   (codacy), test
