+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- PSOLGENT seed handling.
+
 ## [0.0.8] - 15/07/2019
 ### Added
 
