### 2019-11-16 848034b
corrected paper, updated docs #14
