### 2019-12-07 3398c77
Removed jpath example, included cgar #15, updated paper details
