+++ paper/paper.md
@@ -8,11 +8,13 @@ tags:
 authors:
   - name: David Torres Sanchez
     orcid: 0000-0002-2894-9432
-    affiliation: 1
+    affiliation: "1, 2"
 affiliations:
  - name: STOR-i, Lancaster University, UK.
    index: 1
-date: July 2019
+ - name: SINTEF Digital, Mathematics and Cybernetics  
+   index: 2
+date: December 2019
 bibliography: paper.bib
 ---
 
@@ -54,115 +56,12 @@ A key component of a CSP problem is to define a set of resources and a function 
 
 ``cspy`` allows for custom and generic REFs to be used. Hence, allowing for a custom consumption of the resources through the graph.
 
-# Example
 
-Consider the following problem. 
+# Examples
 
-*Example.* Jane is part-time postwoman working in Delft, Netherlands. However, she is assigned a small area (the Indische Buurt neighbourhood) so when planning her daily route she wants to make it as long and exciting as possible. 
-That is, when planning her routes she has to consider the total shift time, sights visited, travel time, and delivery time. Her shift has to be at most 5 hours.
+For some real-world examples, see the repository.
 
-
-This problem can easily be modelled as a CSP problem. 
-With the description above, the set of resources can be defined as,
-
-```python
-R = ['sights', 'shift', 'travel-time', 'delivery-time'] 
-# len(R) = 4
-```
-
-Let ``G`` denote a directed graph with edges to/from all streets of the Indische Buurt neighbourhood. Each edge has an attribute ``weight`` proportional to the distance (in km) between the two nodes and an attribute ``res_cost`` which is an array (specifically, a ``numpy.array``) with length ``len(R)``. To maximise the distance of the path, as required by Jane, we simple negate the distance, hence, solving a shortest path problem with negative distance will be the equivalent to solving a longest path problem. The entries of ``res_cost`` have the same order as the entries in ``R``.
-The first entry of this array, corresponds to the ``'sights'`` resource, i.e. how many sights there are along a specific edge. The last entry of this array, corresponds to the ``'delivery-time'`` resource, i.e. time taken to deliver post along a specific edge. The remaining entries can be initialised to be 0.
-Also, when defining ``G``, one has to specify the number of resources ``n_res``, which also has to be equal to ``len(R)``.
-
-```python
-from networkx import DiGraph
-
-G = DiGraph(directed=True, n_res=4)  # init network
-```
-
-Now let's assume we have a function ``generate_district_network`` that can create the appropriate network.
-
-```python
-from read_input import generate_district_network  
-# function to generate network from data
-
-G = generate_district_network()
-n_edges = len(G.edges())  # number of edges in network
-```
-
-To define the custom REFs,  ``jane_REF``, that controls how resources evolve throughout the path, we require two inputs: an array of current cumulative resource values ``res``, and the edge that is being considered for an extension of a path ``edge`` (which consists of two nodes and the edge data).
-
-```python
-from numpy import array
-
-def jane_REF(res, edge):
-    arr = array(res)  # local array
-    i, j, edge_data = edge[:]  # unpack edge
-    # i, j : string, edge_data : dict
-    # Update 'sights' resource
-    arr[0] += edge_data['res_cost'][0]
-    # Update 'travel-time' resource (distance/speed)
-    arr[2] += - edge_data['weight'] / float(WALKING_SPEED)
-    # Update 'delivery-time' resource
-    arr[3] += edge_data['res_cost'][3]
-    # Update 'shift' resource
-    arr[1] += (arr[2] + arr[3])  # travel-time + delivery-time
-    return arr
[…]
