+++ examples/cgar/README.md
@@ -1,0 +1,7 @@
+# cgar
+
+Column generation for the aircraft recovery problem using ``cspy``.
+
+Please see [cgar.pdf](cgar.pdf) for more details.
+
+Extract from my thesis, all rights reserved.

[…]
