### 2019-12-07 torressa/cspy#15 @torressa
I've included an example for some ongoing work using column generation. Unfortunately, I haven't managed to come up with a simpler example.
