### 2019-12-07 #1655 @poulson
Hi @torressa and @stsievert -- in our previous discussion a few weeks ago it looked like @torressa just needed to finalize some changes requested by @stsievert.

@torressa: How are those changes coming along?

@stsievert: Are you happy with the changes so far?
