### 2019-12-09 #1655 @torressa
Hi @poulson! Thanks for the follow up.
I have made a few changes to accommodate for the comments.
I might still add some additional examples and update the docs.
