### 2019-12-14 torressa/cspy#15 @stsievert
I think the example previously in the paper ("Jane is part-time postwoman working ...") is a great example. I'd love to see that in the documentation.
