### 2019-12-16 torressa/cspy#15 @stsievert
I’m a fan of simple. It’ll lower the barrier to understanding the software.

I’m also a fan of the in-depth airport example. It helps readers who know the math and what they’re doing understand the software, and see some more advanced uses.
