### 2019-12-23 3aeb4c1
Fixed preprocessing no resource found
