### 2019-12-24 575d702
avoid negative cost cycles #17
