+++ cspy/tests/tests.py
@@ -1,147 +1,0 @@
-import sys
-import unittest
-from networkx import DiGraph
-from numpy import array
-from numpy.random import RandomState
-sys.path.append("../")
-from cspy.algorithms.bidirectional import BiDirectional
-from cspy.algorithms.tabu import Tabu
-from cspy.algorithms.greedy_elimination import GreedyElim
-from cspy.algorithms.grasp import GRASP
-from cspy.algorithms.psolgent import PSOLGENT
-from cspy.label import Label
-
-
-class TestsBasic(unittest.TestCase):
-    """ Tests for finding the resource constrained shortest
-    path of simple DiGraph using the BiDirectional algorithm."""
-
-    def setUp(self):
-        self.max_res, self.min_res = [4, 20], [1, 0]
-        # Create simple digraph to test algorithm
-        self.G = DiGraph(directed=True, n_res=2)
-        self.G.add_edge('Source', 'A', res_cost=array([1, 2]), weight=-1)
-        self.G.add_edge('A', 'B', res_cost=array([1, 0.3]), weight=-1)
-        # self.G.add_edge('A', 'C', res_cost=[1, 0.1], weight=1)
-        self.G.add_edge('B', 'C', res_cost=array([1, 3]), weight=-10)
-        self.G.add_edge('B', 'Sink', res_cost=array([1, 2]), weight=10)
-        self.G.add_edge('C', 'Sink', res_cost=array([1, 10]), weight=-1)
-
-        # Create erratic digraph to test exception handling
-        self.E = DiGraph(directed=True)
-        self.E.add_edge('Source', 'A')
-
-        # Create digraph with negative resource costs with unreachable node 'B'
-        self.H = DiGraph(directed=True, n_res=2)
-        self.H.add_edge('Source', 'A', res_cost=array([1, 2]), weight=0)
-        self.H.add_edge('A', 'C', res_cost=array([-1, 0.3]), weight=0)
-        self.H.add_edge('A', 'B', res_cost=array([-1, 3]), weight=0)
-        # Unreachable node B
-        self.H.add_edge('B', 'D', res_cost=array([-1, 2]), weight=0)
-        self.H.add_edge('C', 'D', res_cost=array([1, 0.1]), weight=0)
-        self.H.add_edge('D', 'Sink', res_cost=array([1, 0.1]), weight=0)
-
-        # Create digraph with a resource infeasible minimum cost path
-        self.J = DiGraph(directed=True, n_res=2)
-        self.J.add_edge('Source', 'A', res_cost=array([1, 1]), weight=1)
-        self.J.add_edge('Source', 'B', res_cost=array([1, 1]), weight=1)
-        # Resource infeasible edge
-        self.J.add_edge('Source', 'C', res_cost=array([10, 1]), weight=10)
-        self.J.add_edge('A', 'C', res_cost=array([1, 1]), weight=1)
-        # Resource infeasible edge
-        self.J.add_edge('A', 'E', res_cost=array([10, 1]), weight=10)
-        # Resource infeasible edge
-        self.J.add_edge('A', 'F', res_cost=array([10, 1]), weight=10)
-        self.J.add_edge('B', 'C', res_cost=array([2, 1]), weight=-1)
-        # Resource infeasible edge
-        self.J.add_edge('B', 'F', res_cost=array([10, 1]), weight=10)
-        # Resource infeasible edge
-        self.J.add_edge('B', 'E', res_cost=array([10, 1]), weight=10)
-        self.J.add_edge('C', 'D', res_cost=array([1, 1]), weight=-1)
-        self.J.add_edge('D', 'E', res_cost=array([1, 1]), weight=1)
-        self.J.add_edge('D', 'F', res_cost=array([1, 1]), weight=1)
-        # Resource infeasible edge
[…]
