+++ cspy/tests/tests_greedyelim.py
@@ -1,0 +1,1 @@
+# TODO Write tests

[…]
