+++ cspy/algorithms/bidirectional.py
@@ -1,16 +1,17 @@
 from __future__ import absolute_import
-from __future__ import print_function
 
 import random
-import logging
-import numpy as np
-from itertools import repeat
 from collections import OrderedDict
+from itertools import repeat
+from logging import getLogger
+
+import numpy as np
+
 # Local module imports
 from cspy.label import Label
 from cspy.preprocessing import check_and_preprocess
 
-log = logging.getLogger(__name__)
+log = getLogger(__name__)
 
 
 class BiDirectional:
@@ -45,7 +46,7 @@ class BiDirectional:
 
     direction : string, optional
         preferred search direction.
-        Either 'both', 'forward', or, 'backward'. Default : 'both'.
+        Either "both", "forward", or, "backward". Default : "both".
 
     .. _REFs : https://cspy.readthedocs.io/en/latest/how_to.html#refs
 
@@ -60,28 +61,16 @@ class BiDirectional:
     :math:`\geq 2`. The edges in the graph must all have a ``res_cost``
     attribute.
 
-    According to the inputs, four different algorithms can be implemented,
-    if you'd like to check which algorithm your running given your resource
-    limits, run ``.name_algorithm(U, L)`` for a log with the classification.
-    If ``direction`` is not given the absolute resource limits have to be
-    given:
+    According to the inputs, four different algorithms can be used.
+    If you'd like to check which algorithm is running given the resource
+    limits, call the method :func:`BiDirectional.name_algorithm()`
+    for a log with the classification.
 
-    :func:`BiDirectional.name_algorithm(U, L)`
-
-    U : float, optional
-        Upper bound for monotone resource
-        (for algorithm classification purposes see Notes).
-
-    L : float, optional
-        Lower bound for monotone resource
-        (for algorithm classification purposes see Notes).
-
-    According to these, we have,
-
-    -  :math:`H_F = H_B > U` or ``direction`` = 'forward': Monodirectional forward labeling algorithm
-    -  :math:`L < H_F = H_B < U`: Bidirectional labeling algorithm with static halfway point
-    -  :math:`H_F = H_B < L` or ``direction`` = 'backward': Monodirectional backward labeling algorithm
-    -  :math:`U = H_F > H_B = L`: Bidirectional labeling algorithm with dynamic halfway point.
+    - ``direction`` = "forward": Monodirectional forward labeling algorithm
+    - :math:`H_F == H_B`: Bidirectional labeling algorithm with static halfway point.
+    - ``direction`` = "backward": Monodirectional backward labeling algorithm
+    - :math:`H_F > H_B`: Bidirectional labeling algorithm with dynamic halfway point.
+    - :math:`H_F < H_B`: The algorithm won't go anywhere!
 
     Example
     -------
@@ -94,16 +83,16 @@ class BiDirectional:
         >>> from networkx import DiGraph
         >>> from numpy import array
         >>> G = DiGraph(directed=True, n_res=2)
-        >>> G.add_edge('Source', 'A', res_cost=array([1, 2]), weight=0)
-        >>> G.add_edge('A', 'B', res_cost=array([1, 0.3]), weight=0)
-        >>> G.add_edge('A', 'C', res_cost=array([1, 0.1]), weight=0)
-        >>> G.add_edge('B', 'C', res_cost=array([1, 3]), weight=-10)
-        >>> G.add_edge('B', 'Sink', res_cost=array([1, 2]), weight=10)
-        >>> G.add_edge('C', 'Sink', res_cost=array([1, 10]), weight=0)
+        >>> G.add_edge("Source", "A", res_cost=array([1, 2]), weight=0)
[…]

[…]
