### 2019-12-27 #1655 @poulson
Thank you, @torressa!

:wave: @stsievert -- Do you have have any lingering concerns given David's updates?

It looks like you might still be concerned about: functionality, performance, examles, statement of need, and community guidelines.
