### 2020-01-08 #1655 @stsievert
I think most of my concerns have been resolved (but will need to look more carefully). I'd still like to see a simple use case on the documentation, alongside a clear description of the features of a CSP problem.
