### 2020-01-23 #1655 @poulson
Hi @hausen :wave: -- please let me know if you think you will be able to make progress on your review in the next few days. @stsievert has been nearly done with theirs for a few weeks and it would be good for us to promptly respond to the authors.
