### 2020-02-07 torressa/cspy#18 @torressa
JOSS REVIEW. Corrections
@torressa @poulson I only found a couple of small issues:

 - [x] **Installation instructions:** `pip install cspy` does not automatically install the dependencies (tried on a clean pyenv)
 - [x] **Example usage:** a couple of the examples in the API docs have errors.
  * [BiDirectional](https://cspy.readthedocs.io/en/latest/api/cspy.BiDirectional.html#example): replace `import cspy` with `from cspy import BiDirectional`
  * [GreedyElim](https://cspy.readthedocs.io/en/latest/api/cspy.GreedyElim.html#example): add `from networkx import DiGraph` line, replace `G = nx.DiGraph(directed=True, n_res=2)` with `G = DiGraph(directed=True, n_res=2)`
  * [PSOLGENT](https://cspy.readthedocs.io/en/latest/api/cspy.PSOLGENT.html#example): replace `n_nodes = len(self.J.nodes())` with `n_nodes = len(G.nodes())`
  * [cgar/requirements.txt](https://github.com/torressa/cspy/blob/master/examples/cgar/requirements.txt): replace `cspy>=0.8` with `cspy>=0.0.8`. Unfortunately, I couldn't test it because it requires a proprietary package, `gurobipy`. @torressa is it possible to eliminate that requirement, or provide another real-world example that doesn't need that package?
 - [x] **README.md**: `Inrich and Desaulniers (2005)` should link to https://www.researchgate.net/publication/227142556_Shortest_Path_Problems_with_Resource_Constraints .

_Originally posted by @hausen in https://github.com/openjournals/joss-reviews/issues/1655#issuecomment-581737265_
