### 2020-02-08 4f68249
Docs and example fixes #18
