+++ examples/cgar/__init__.py
@@ -1,6 +1,10 @@
 from read_input import read_input
 from algorithm import algorithm
 
+import logging
+
+logging.basicConfig(level=logging.INFO)
+
 airline = 'Finnair'
 
 DataObj = read_input(airline=airline)
