+++ cspy/algorithms/greedy_elimination.py
@@ -70,7 +70,8 @@ class GreedyElim:
     .. code-block:: python
 
         >>> from cspy import GreedyElim
-        >>> G = nx.DiGraph(directed=True, n_res=2)
+        >>> from networkx import DiGraph
+        >>> G = DiGraph(directed=True, n_res=2)
         >>> G.add_edge('Source', 'A', res_cost=[1, 1], weight=1)
         >>> G.add_edge('Source', 'B', res_cost=[1, 1], weight=1)
         >>> G.add_edge('A', 'C', res_cost=[1, 1], weight=1)
