+++ examples/cgar/requirements.txt
@@ -1,5 +1,5 @@
 networkx>=2.2
 pandas>=0.24.2
 numpy>=1.16.2
-cspy>=0.8
+cspy>=0.0.8
 gurobipy>=5.0.2
