### 2020-02-08 c4b3787
translated column generation example into pulp #18
