+++ cspy/tests/tests_tabu.py
@@ -10,7 +10,7 @@ from cspy.algorithms.tabu import Tabu
 
 class TestsTabu(unittest.TestCase):
     """ Tests for finding the resource constrained shortest
-    path of simple DiGraph using the BiDirectional algorithm."""
+    path of simple DiGraph using the Tabu algorithm."""
 
     def setUp(self):
         self.max_res, self.min_res = [5, 5], [0, 0]
