+++ cspy/tests/tests_grasp.py
@@ -10,7 +10,7 @@ from cspy.algorithms.grasp import GRASP
 
 class TestsGRASP(unittest.TestCase):
     """ Tests for finding the resource constrained shortest
-    path of simple DiGraph using the BiDirectional algorithm."""
+    path of simple DiGraph using the GRASP algorithm."""
 
     def setUp(self):
         self.max_res, self.min_res = [5, 5], [0, 0]
