+++ .circleci/config.yml
@@ -42,9 +42,8 @@ jobs:
           command: |
             python3 -m venv venv
             . venv/bin/activate
-            python3 -m unittest
             cd ~/cspy/cspy
-            ls
+            python3 -m unittest
             python3 -m coverage run -m unittest
             codecov
 
