### 2020-02-09 #1655 @torressa
Sorted!
1. Updated the pypi package (bumped to 0.0.10) so that it would automatically install dependencies.
2. Corrected mistakes in doc examples. Also translated airport example to use PuLP instead of gurobipy.
3. Fixed! (Together with all the other links that weren't working)
