### 2020-03-06 886cf9f
Simplified BiDirectional #17 fixed Tabu heuristic function #20
