### 2020-03-07 af43c66
implemented half-way procedure #21 fixed testing errors
