+++ cspy/algorithms/bidirectional.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 
 from itertools import repeat
+from operator import add, sub
 from logging import getLogger
 from collections import OrderedDict, deque
 
@@ -144,6 +145,9 @@ class BiDirectional:
         if REF_forward and REF_backward:
             Label._REF_forward = REF_forward
             Label._REF_backward = REF_backward
+        else:
+            Label._REF_forward = add
+            Label._REF_backward = sub
         # Init with seed if given
         if seed is None:
             self.random_state = RandomState()
@@ -152,9 +156,7 @@ class BiDirectional:
         elif isinstance(seed, RandomState):
             self.random_state = seed
         else:
-            raise Exception(
-                '{} cannot be used to seed numpy.random.RandomState'.format(
-                    seed))
+            raise Exception("{} cannot be used to seed".format(seed))
 
     def run(self):
         while self.currentLabel["forward"] or self.currentLabel["backward"]:
@@ -162,10 +164,9 @@ class BiDirectional:
             if direc:
                 self._algorithm(direc)
                 self._check_dominance(direc)
-            elif not direc or self._terminate(direc):
+            else:
                 break
-        # Otherwise return either path
-        return self._join_paths()
+        return self._process_paths()
 
     ###########################
     # Classify Algorithm Type #
@@ -237,10 +238,12 @@ class BiDirectional:
         list(map(self._propagate_label, edges, repeat(direc, len(edges))))
         # Extend label
         next_label = self._get_next_label(direc)
+        # If next label already been processed get another one.
         if next_label in self.processedLabels[direc]:
             next_label = self._get_next_label(direc, next_label)
         # Update current label
         self.currentLabel[direc] = next_label
+        log.debug("current label = {}".format(next_label))
 
     def _propagate_label(self, edge, direc):
         # Label propagation #
@@ -345,74 +348,102 @@ class BiDirectional:
     def _remove_unprocessed_labels(self, keys_to_pop, direc):
         # Remove all processed labels from unprocessed dict
         for key_to_pop in list(set(keys_to_pop)):
-            if key_to_pop in self.unprocessedLabels[direc]:
-                log.debug("Key {} removed".format(key_to_pop))
-                del self.unprocessedLabels[direc][key_to_pop]
             for k, sub_dict in self.unprocessedLabels[direc].items():
                 if key_to_pop in sub_dict:
                     _idx = sub_dict.index(key_to_pop)
                     log.debug("Key {} removed from sub_dict".format(key_to_pop))
                     del self.unprocessedLabels[direc][k][_idx]
 
-    ###############
-    # TERMINATION #
-    ###############
-    def _terminate(self, direc):
+    ###################
+    # PATH PROCESSING #
+    ###################
+    def _process_paths(self):
+        # Processing of output path.
         if self.direc_in == "both":
-            if (not self.unprocessedLabels["forward"] or
-                    not self.unprocessedLabels["backward"]):
-                return True
-        else:
-            if (self.direc_in == "forward" and
-                    self.finalLabel["forward"].path[-1] == "Sink" and
-                    not self.unprocessedLabels["forward"]):
-                return True
-            elif (self.direc_in == "backward" and
-                  self.finalLabel["backward"].path[-1] == "Source" and
-                  not self.unprocessedLabels["backward"]):
-                return True
-
-    #################
-    # PATH CHECKING #
-    #################
-    def _join_paths(self):
-        # check if paths are eligible to be joined
-        if self.direc_in == "both":
+            # If bi-directional algorithm used, run halfway procedure.
             return self._check_paths()
         else:
[…]
