### 2020-03-09 #1655 @arfon
@hausen - can you try accepting the invite at this page which should allow you to update the checklist https://github.com/openjournals/joss-reviews/invitations
