### 2020-03-11 #1655 @poulson
Thank you, Rodrigo (@hausen)!

Hi Scott (@stsievert) :wave: -- it seems you still have four boxes unchecked:
 - Statement of need
 - Example Usage
 - Functionality Documentation
 - References

Do you have any lingering concerns on any of these four?
