### 2020-03-12 2896d33
changed label dominance and final path saving #22
