+++ cspy/algorithms/bidirectional.py
@@ -18,9 +18,12 @@ log = getLogger(__name__)
 class BiDirectional:
     """
     Implementation of the bidirectional labeling algorithm with dynamic
-    half-way point (`Tilk 2017`_).
+    half-way point (`Tilk 2017`_). 
+    This requires the half-way procedure from `Righini and Salani (2006)`_,
+    also implemented.
     Depending on the range of values for U, L, we get
     four different algorithms. See self.name_algorithm and Notes.
+
 
     Parameters
     ----------
@@ -100,8 +103,8 @@ class BiDirectional:
         ["Source", "A", "B", "C", "Sink"]
 
     .. _Tilk 2017: https://www.sciencedirect.com/science/article/pii/S0377221717302035
+    .. _Righini and Salani (2006): https://www.sciencedirect.com/science/article/pii/S1572528606000417
     """
-
     def __init__(self,
                  G,
                  max_res,
@@ -191,20 +194,20 @@ class BiDirectional:
     #############
     def _get_direction(self):
         if self.direc_in == "both":
-            if (self.currentLabel["forward"] and
-                    not self.currentLabel["backward"]):
+            if (self.currentLabel["forward"]
+                    and not self.currentLabel["backward"]):
                 return "forward"
-            elif (not self.currentLabel["forward"] and
-                  self.currentLabel["backward"]):
+            elif (not self.currentLabel["forward"]
+                  and self.currentLabel["backward"]):
                 return "backward"
-            elif (self.currentLabel["forward"] and
-                  self.currentLabel["backward"]):
+            elif (self.currentLabel["forward"]
+                  and self.currentLabel["backward"]):
                 return self.random_state.choice(["forward", "backward"])
             else:  # if both are empty
                 return
         else:
-            if (not self.currentLabel["forward"] and
-                    not self.currentLabel["backward"]):
+            if (not self.currentLabel["forward"]
+                    and not self.currentLabel["backward"]):
                 return
             elif not self.currentLabel[self.direc_in]:
                 return
@@ -263,15 +266,15 @@ class BiDirectional:
             # Try to get labels create from the current one.
             _labels = list(lab for lab in self.unprocessedLabels[direc][
                 self.currentLabel[direc]]
-                           if (lab not in self.processedLabels[direc] and
-                               lab != exclude_label))
+                           if (lab not in self.processedLabels[direc]
+                               and lab != exclude_label))
             return min(_labels, key=lambda x: x.weight)
         except ValueError:
             # No more keys to be processed under current label
             self.processedLabels[direc].append(self.currentLabel[direc])
             _labels = list(lab for lab in self.unprocessedLabels[direc].keys()
-                           if (lab not in self.processedLabels[direc] and
-                               lab != exclude_label))
+                           if (lab not in self.processedLabels[direc]
+                               and lab != exclude_label))
             if _labels:
                 return min(_labels, key=lambda x: x.weight)
             else:
@@ -281,34 +284,37 @@ class BiDirectional:
             return None
 
     def _save_current_best_label(self, direc):
+        current_label = self.currentLabel[direc]
+        final_label = self.finalLabel[direc]
         try:
-            if self.currentLabel[direc].dominates(self.finalLabel[direc],
-                                                  direc):
+            current_label_dominates = current_label.dominates(
+                final_label, direc)
+            final_label_dominates = final_label.dominates(current_label, direc)
+            # Current label dominates final
+            if current_label_dominates:
                 log.debug("Saving {} as best, with path {}".format(
-                    self.currentLabel[direc], self.currentLabel[direc].path))
-                self.finalLabel[direc] = self.currentLabel[direc]
[…]
