+++ .circleci/config.yml
@@ -2,25 +2,22 @@
 #
 # Check https://circleci.com/docs/2.0/language-python/ for more details
 #
-version: 2.1
+version: 2.0
+
 orbs:
   codecov: codecov/codecov@1.0.0
 jobs:
-  build:
+  "python-3.8":
     docker:
-      - image: circleci/python:3.6.1
-
+      - image: circleci/python:3.8
     working_directory: ~/cspy
-
     steps:
       - checkout
-
       - restore_cache:
           keys:
             - v1-dependencies-{{ checksum "requirements.txt" }}
             # fallback to using the latest cache if no exact match is found
             - v1-dependencies-
-
       - run:
           name: install dependencies
           command: |
@@ -31,12 +28,82 @@ jobs:
       - run:
           name: codecov-install
           command: sudo pip install codecov
-
       - save_cache:
           paths:
             - ./venv
           key: v1-dependencies-{{ checksum "requirements.txt" }}
-
+      - run:
+          name: run tests
+          command: |
+            python3 -m venv venv
+            . venv/bin/activate
+            cd ~/cspy/
+            python3 -m unittest
+            codecov
+      - store_artifacts:
+          path: test-reports
+          destination: test-reports
+  "python-3.7":
+    docker:
+      - image: circleci/python:3.7
+    working_directory: ~/cspy
+    steps:
+      - checkout
+      - restore_cache:
+          keys:
+            - v1-dependencies-{{ checksum "requirements.txt" }}
+            # fallback to using the latest cache if no exact match is found
+            - v1-dependencies-
+      - run:
+          name: install dependencies
+          command: |
+            python3 -m venv venv
+            . venv/bin/activate
+            pip install -r requirements.txt
+            pip install -r examples/cgar/requirements.txt
+      - run:
+          name: codecov-install
+          command: sudo pip install codecov
+      - save_cache:
+          paths:
+            - ./venv
+          key: v1-dependencies-{{ checksum "requirements.txt" }}
+      - run:
+          name: run tests
+          command: |
+            python3 -m venv venv
+            . venv/bin/activate
+            cd ~/cspy/
+            python3 -m unittest
+            codecov
+      - store_artifacts:
+          path: test-reports
+          destination: test-reports
+  "python-3.6":
+    docker:
+      - image: circleci/python:3.6
+    working_directory: ~/cspy
+    steps:
+      - checkout
+      - restore_cache:
+          keys:
+            - v1-dependencies-{{ checksum "requirements.txt" }}
+            # fallback to using the latest cache if no exact match is found
+            - v1-dependencies-
+      - run:
+          name: install dependencies
+          command: |
+            python3 -m venv venv
+            . venv/bin/activate
+            pip install -r requirements.txt
+            pip install -r examples/cgar/requirements.txt
+      - run:
+          name: codecov-install
+          command: sudo pip install codecov
+      - save_cache:
+          paths:
+            - ./venv
+          key: v1-dependencies-{{ checksum "requirements.txt" }}
       - run:
           name: run tests
           command: |
@@ -44,8 +111,86 @@ jobs:
             . venv/bin/activate
             cd ~/cspy/
             python3 -m coverage run -m unittest
-            codecov
-
       - store_artifacts:
           path: test-reports
           destination: test-reports
+  "python-3.5":
+    docker:
+      - image: circleci/python:3.5
+    working_directory: ~/cspy
+    steps:
+      - checkout
+      - restore_cache:
+          keys:
+            - v1-dependencies-{{ checksum "requirements.txt" }}
+            # fallback to using the latest cache if no exact match is found
+            - v1-dependencies-
+      - run:
+          name: install dependencies
+          command: |
+            python3 -m venv venv
[…]
