### 2020-03-12 98026cb
reduced pandas version for python-3.4
