+++ examples/cgar/requirements.txt
@@ -1,5 +1,5 @@
 networkx>=2.2
-pandas>=0.24.2
+pandas>=0.22
 PuLP>=2.0
 numpy>=1.16.2
 cspy>=0.0.9
