### 2020-03-12 5ada456
removed python3.4 as collections.deque is incompatible
