+++ examples/cgar/requirements.txt
@@ -2,4 +2,3 @@ networkx>=2.2
 pandas>=0.22
 PuLP>=2.0
 numpy>=1.16.2
-cspy>=0.0.9
