+++ .circleci/config.yml
@@ -139,41 +139,6 @@ jobs:
             . venv/bin/activate
             cd ~/cspy/
             python3 -m unittest
-  "python-3.4":
-    docker:
-      - image: circleci/python:3.4
-    working_directory: ~/cspy
-    steps:
-      - checkout
-      - restore_cache:
-          keys:
-            - v1-dependencies-{{ checksum "requirements.txt" }}
-            # fallback to using the latest cache if no exact match is found
-            - v1-dependencies-
-      - run:
-          name: install dependencies
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            pip install -r requirements.txt
-            pip install -r examples/cgar/requirements.txt
-      - run:
-          name: codecov-install
-          command: sudo pip install codecov
-      - save_cache:
-          paths:
-            - ./venv
-          key: v1-dependencies-{{ checksum "requirements.txt" }}
-      - run:
-          name: run tests
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            cd ~/cspy/
-            python3 -m unittest
-      - store_artifacts:
-          path: test-reports
-          destination: test-reports
 
 workflows:
   version: 2
@@ -183,4 +148,3 @@ workflows:
       - "python-3.7"
       - "python-3.6"
       - "python-3.5"
-      - "python-3.4"
