### 2020-03-12 a2f54ba
updated setup and readme
