+++ setup.py
@@ -2,7 +2,7 @@ import setuptools
 
 setuptools.setup(
     name="cspy",
-    version="0.0.2",
+    version="0.0.12",
     description=
     "A collection of algorithms for the (Resource) Constrained Shortest Path Problem",
     license="MIT",
@@ -12,12 +12,16 @@ setuptools.setup(
         "shortest path", "resource constrained shortest path",
         "bidirectional algorithm", "tabu search"
     ],
-    long_description_content_type="text/markdown",
-    long_description=open("README.md", "r").read(),
+    long_description_content_type="text/x-rst",
+    long_description=open("README.rst", "r").read(),
     url="https://github.com/torressa/cspy",
     packages=setuptools.find_packages(),
+    install_requires=['networkx', 'numpy'],
     classifiers=[
-        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.5",
+        "Programming Language :: Python :: 3.6",
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
