+++ CHANGELOG.md
@@ -9,13 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Documentation.
-- BiDirectional algorithm:
-	- **Removed** termination criteria.
-	- Implemented half way procedure from [Righini and Salani (2006)](https://www.sciencedirect.com/science/article/pii/S1572528606000417) in ``self._half_way`` (Closes #21).
-	- Changed label dominance to an equivalent but more elegant function.
-	- Changed final label saving to account for when two label do not dominate each other.
+- BiDirectional algorithm: - **Removed** termination criteria. - Implemented half way procedure from [Righini and Salani (2006)](https://www.sciencedirect.com/science/article/pii/S1572528606000417) in `self._half_way` (Closes #21). - Changed label dominance to an equivalent but more elegant function. - Changed final label saving to account for when two label do not dominate each other.
 
-## [0.0.11] 06/03/2020
+## [v0.0.11] 06/03/2020
 
 ## Fixed
 
@@ -25,24 +21,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Documentation.
-- BiDirectional algorithm:
-	- Final label comparisons.
-	- Seed handling for testing.
-	- Renamed variables to avoid confusion.
-	- Avoiding getting stuck processing cycles of input graphs.
-	- Ensuring that edges in path correspond to an edge in the input graph.
-	- Avoid overwriting inputs (`max_res` and `min_res`).
-   	- Removed loops in `_get_next_label` and `_check_dominance` in favour of list comprehensions.
-	- Use of `collections`.
-	- logs for debugging in BiDirectional.
-	- added `_save_current_best_label`.
-	- Changed type of `self.finalLabel["direction"]` from list to `Label`. 
-	- **Removed** termination criteria.
-	- Implemented half way procedure from [Righini and Salani (2006)](https://www.sciencedirect.com/science/article/pii/S1572528606000417) in ``self._half_way`` (Closes #21).
+- BiDirectional algorithm: - Final label comparisons. - Seed handling for testing. - Renamed variables to avoid confusion. - Avoiding getting stuck processing cycles of input graphs. - Ensuring that edges in path correspond to an edge in the input graph. - Avoid overwriting inputs (`max_res` and `min_res`). - Removed loops in `_get_next_label` and `_check_dominance` in favour of list comprehensions. - Use of `collections`. - logs for debugging in BiDirectional. - added `_save_current_best_label`. - Changed type of `self.finalLabel["direction"]` from list to `Label`.
 
 - Re-organised. Moved `label.py` and `path.py` into `algorithms/`.
 
-## [0.0.10] - 09/02/2020
+## [v0.0.10] - 09/02/2020
 
 ## Added
 
@@ -51,10 +34,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Changed
 
 - Documentation.
-- Translated ``examples/cgar`` from gurobipy to pulp.
+- Translated `examples/cgar` from gurobipy to pulp.
 - CI build.
 
-## [0.0.9] - 25/12/2019
+## [v0.0.9] - 25/12/2019
 
[…]
