### 2020-03-12 ef3e966
added features and updated docs
