+++ tests/tests_tabu.py
@@ -9,9 +9,10 @@ from cspy.algorithms.tabu import Tabu
 
 
 class TestsTabu(unittest.TestCase):
-    """ Tests for finding the resource constrained shortest
-    path of simple DiGraph using the Tabu algorithm."""
-
+    """
+    Tests for finding the resource constrained shortest
+    path of simple DiGraph using the Tabu algorithm.
+    """
     def setUp(self):
         self.max_res, self.min_res = [5, 5], [0, 0]
         # Create digraph with a resource infeasible minimum cost path
@@ -40,8 +41,14 @@ class TestsTabu(unittest.TestCase):
         self.G.add_edge('E', 'Sink', res_cost=array([1, 1]), weight=1)
 
     def testTabu(self):
-        path = Tabu(self.G, self.max_res, self.min_res).run()
+        tabu = Tabu(self.G, self.max_res, self.min_res)
+        tabu.run()
+        path = tabu.path
+        cost = tabu.total_cost
+        total_res = tabu.consumed_resources
         self.assertEqual(path, ['Source', 'A', 'C', 'D', 'E', 'Sink'])
+        self.assertEqual(cost, 3)
+        self.assertTrue(all(total_res == [5, 5]))
 
     def testInputExceptions(self):
         # Check whether wrong input raises exceptions
