### 2020-03-14 #1655 @stsievert
> Do you have any lingering concerns on any of these four?

- Problem statement: no. I've checked the box.
- [ ] Example usage: yes. The link to "examples" is broken on the docs (on *[Using cspy]*), and I'd like to see a simple example as expressed in https://github.com/torressa/cspy/issues/15#issuecomment-566115714
- [ ] Functionality Documentation: one small concerns. The `n_res` parameter on the algorithms page is not defined (e.g, [GRASP]). It is explained on *[Using cspy]*, but confusing when I first see an algorithm. Perhaps if I were a domain expert I would realize.
- References: no.

I've checked the boxes above.

[Using cspy]:https://cspy.readthedocs.io/en/latest/how_to.html
[GRASP]:https://cspy.readthedocs.io/en/latest/api/cspy.GRASP.html
