### 2020-03-14 162415d
Split preprocessing and checking, added resource inversion for backward paths #23
