+++ cspy/algorithms/grasp.py
@@ -10,7 +10,10 @@ from numpy.random import choice
 from random import sample, randint
 from cspy.algorithms.path import Path
 from itertools import permutations, repeat
-from cspy.preprocessing import check_and_preprocess
+
+# Local imports
+from cspy.checking import check
+from cspy.preprocessing import preprocess_graph
 
 log = logging.getLogger(__name__)
 
@@ -117,10 +120,10 @@ class GRASP:
                  max_iter=100,
                  max_localiter=10,
                  alpha=0.2):
-        # Check input graph and parameters
-        G = check_and_preprocess(preprocess, G, max_res, min_res, REF)
-        # Input parameters
-        self.G = G
+        # Check inputs
+        check(G, max_res, min_res, REF)
+        # Preprocess graph
+        self.G = preprocess_graph(G, max_res, min_res, preprocess, REF)
         self.max_res = max_res
         self.min_res = min_res
         self.max_iter = max_iter
@@ -155,6 +158,8 @@ class GRASP:
         """
         Get list with nodes in calculated path.
         """
+        if not self.best_path:
+            raise Exception("Please call the .run() method first")
         return self.best_path.path
 
     @property
@@ -162,6 +167,8 @@ class GRASP:
         """
         Get accumulated cost along the path.
         """
+        if not self.best_path:
+            raise Exception("Please call the .run() method first")
         return self.best_path.cost
 
     @property
@@ -169,6 +176,8 @@ class GRASP:
         """
         Get accumulated resources consumed along the path.
         """
+        if not self.best_path:
+            raise Exception("Please call the .run() method first")
         return self.best_path.total_res
 
     def _algorithm(self):
