+++ cspy/algorithms/psolgent.py
@@ -13,18 +13,20 @@ from numpy.random import RandomState
 from numpy import (argmin, array, copy, diag_indices_from, exp, dot, zeros,
                    ones, where)
 
+# Local imports
+from cspy.checking import check
 from cspy.algorithms.path import Path
-from cspy.preprocessing import check_and_preprocess
+from cspy.preprocessing import preprocess_graph
 
 log = getLogger(__name__)
 
 
 class StandardGraph:
     def __init__(self, G, max_res, min_res, REF, preprocess):
-        # Check input graph and parameters
-        G = check_and_preprocess(preprocess, G, max_res, min_res, REF)
-        # Input parameters
-        self.G = G
+        # Check inputs
+        check(G, max_res, min_res, REF)
+        # Preprocess graph
+        self.G = preprocess_graph(G, max_res, min_res, preprocess, REF)
         self.max_res = max_res
         self.min_res = min_res
         self.path_list = None
@@ -287,6 +289,8 @@ class PSOLGENT(StandardGraph):
         """
         Get list with nodes in calculated path.
         """
+        if not self.best_path:
+            raise Exception("Please call the .run() method first")
         return self.best_path.path
 
     @property
@@ -294,6 +298,8 @@ class PSOLGENT(StandardGraph):
         """
         Get accumulated cost along the path.
         """
+        if not self.best_path:
+            raise Exception("Please call the .run() method first")
         return self.best_path.cost
 
     @property
@@ -301,6 +307,8 @@ class PSOLGENT(StandardGraph):
         """
         Get accumulated resources consumed along the path.
         """
+        if not self.best_path:
+            raise Exception("Please call the .run() method first")
         return self.best_path.total_res
 
     def _init_swarm(self):
