+++ cspy/__init__.py
@@ -3,13 +3,14 @@ from cspy.algorithms.tabu import Tabu
 from cspy.algorithms.greedy_elimination import GreedyElim
 from cspy.algorithms.psolgent import PSOLGENT
 from cspy.algorithms.grasp import GRASP
-from cspy.preprocessing import check_and_preprocess
+from cspy.checking import check
+from cspy.preprocessing import preprocess_graph
 
 name = "cspy"
 
 __all__ = [
-    'BiDirectional', 'Tabu', 'GreedyElim', 'PSOLGENT', 'GRASP',
-    'check_and_preprocess'
+    'BiDirectional', 'Tabu', 'GreedyElim', 'PSOLGENT', 'GRASP', 'check',
+    'preprocess_graph'
 ]
 
 __version__ = '0.0.10'
