+++ setup.py
@@ -18,7 +18,6 @@ setuptools.setup(
     packages=setuptools.find_packages(),
     install_requires=['networkx', 'numpy'],
     classifiers=[
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
