+++ .circleci/config.yml
@@ -13,11 +13,6 @@ jobs:
     working_directory: ~/cspy
     steps:
       - checkout
-      - restore_cache:
-          keys:
-            - v1-dependencies-{{ checksum "requirements.txt" }}
-            # fallback to using the latest cache if no exact match is found
-            - v1-dependencies-
       - run:
           name: install dependencies
           command: |
@@ -25,13 +20,6 @@ jobs:
             . venv/bin/activate
             pip install -r requirements.txt
             pip install -r examples/cgar/requirements.txt
-      - run:
-          name: codecov-install
-          command: sudo pip install codecov
-      - save_cache:
-          paths:
-            - ./venv
-          key: v1-dependencies-{{ checksum "requirements.txt" }}
       - run:
           name: run tests
           command: |
@@ -45,11 +33,6 @@ jobs:
     working_directory: ~/cspy
     steps:
       - checkout
-      - restore_cache:
-          keys:
-            - v1-dependencies-{{ checksum "requirements.txt" }}
-            # fallback to using the latest cache if no exact match is found
-            - v1-dependencies-
       - run:
           name: install dependencies
           command: |
@@ -57,13 +40,6 @@ jobs:
             . venv/bin/activate
             pip install -r requirements.txt
             pip install -r examples/cgar/requirements.txt
-      - run:
-          name: codecov-install
-          command: sudo pip install codecov
-      - save_cache:
-          paths:
-            - ./venv
-          key: v1-dependencies-{{ checksum "requirements.txt" }}
       - run:
           name: run tests
           command: |
@@ -77,11 +53,6 @@ jobs:
     working_directory: ~/cspy
     steps:
       - checkout
-      - restore_cache:
-          keys:
-            - v1-dependencies-{{ checksum "requirements.txt" }}
-            # fallback to using the latest cache if no exact match is found
-            - v1-dependencies-
       - run:
           name: install dependencies
           command: |
@@ -89,15 +60,9 @@ jobs:
             . venv/bin/activate
             pip install -r requirements.txt
             pip install -r examples/cgar/requirements.txt
+            pip install codecov
       - run:
-          name: codecov-install
-          command: sudo pip install codecov
-      - save_cache:
-          paths:
-            - ./venv
-          key: v1-dependencies-{{ checksum "requirements.txt" }}
-      - run:
-          name: run tests
+          name: run tests and upload coverage
           command: |
             python3 -m venv venv
             . venv/bin/activate
@@ -107,38 +72,6 @@ jobs:
       - store_artifacts:
           path: test-reports
           destination: test-reports
-  "python-3.5":
-    docker:
-      - image: circleci/python:3.5
-    working_directory: ~/cspy
-    steps:
-      - checkout
-      - restore_cache:
-          keys:
-            - v1-dependencies-{{ checksum "requirements.txt" }}
-            # fallback to using the latest cache if no exact match is found
-            - v1-dependencies-
-      - run:
-          name: install dependencies
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            pip install -r requirements.txt
-            pip install -r examples/cgar/requirements.txt
-      - run:
-          name: codecov-install
-          command: sudo pip install codecov
-      - save_cache:
-          paths:
-            - ./venv
-          key: v1-dependencies-{{ checksum "requirements.txt" }}
-      - run:
-          name: run tests
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            cd ~/cspy/
-            python3 -m unittest
 
 workflows:
   version: 2
@@ -147,4 +80,3 @@ workflows:
       - "python-3.8"
       - "python-3.7"
       - "python-3.6"
-      - "python-3.5"
