### 2020-03-14 0430cb3
Codacy pull request fixes #24
