### 2020-03-14 c1ff9b6
More codacy fixes #24
