+++ cspy/__init__.py
@@ -13,4 +13,4 @@ __all__ = [
     'preprocess_graph'
 ]
 
-__version__ = '0.0.10'
+__version__ = '0.0.12'
