+++ docs/api/cspy.preprocess_graph.rst
@@ -2,4 +2,3 @@ cspy.preprocess_graph
 =====================
 
 .. autofunction:: cspy.preprocess_graph
-	
