### 2020-03-14 b339090
Docs fix, updated compare in changelog and badge
