+++ docs/api/cspy.Label.rst
@@ -1,5 +1,0 @@
-cspy.Label
-==========
-
-.. autoclass:: cspy.Label
-	
