+++ docs/index.rst
@@ -56,7 +56,6 @@ and, increased efficiency
    api/cspy.PSOLGENT
    api/cspy.preprocess_graph
    api/cspy.check
-   api/cspy.Label
 
     
     
