### 2020-03-14 9017995
Updated set up for readthedocs build
