### 2020-03-16 torressa/cspy#15 @torressa
Reopened, add simple example.
https://github.com/openjournals/joss-reviews/issues/1655#issuecomment-598995867
