### 2020-03-23 3416d53
Clean up bidirectional, joining alg #25
