+++ cspy/algorithms/bidirectional.py
@@ -1,11 +1,12 @@
 from __future__ import absolute_import
 
+from copy import deepcopy
 from itertools import repeat
 from operator import add, sub
 from logging import getLogger
 from collections import OrderedDict, deque
 
-from numpy import zeros
+from numpy import zeros, array
 from numpy.random import RandomState
 
 # Local module imports
@@ -119,9 +120,11 @@ class BiDirectional:
                                   REF_backward)
         self.direc_in = direction
         self.max_res, self.min_res = max_res.copy(), min_res.copy()
-        self.max_res_in, self.min_res_in = max_res.copy(), min_res.copy()
+        self.max_res_in, self.min_res_in = array(max_res.copy()), array(
+            min_res.copy())
         # To expose results
         self.best_label = None
+
         # Algorithm specific parameters #
         # Current forward and backward labels
         self.currentLabel = OrderedDict({
@@ -132,19 +135,16 @@ class BiDirectional:
         })
         # Unprocessed labels dict (both directions)
         self.unprocessedLabels = OrderedDict({
-            "forward": OrderedDict({}),
-            "backward": OrderedDict({})
+            "forward": deque(),
+            "backward": deque()
         })
         # Processed labels dict
         self.processedLabels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
-        # Final labels dict.
-        self.finalLabel = OrderedDict({
-            "forward": self.currentLabel["forward"],
-            "backward": self.currentLabel["backward"]
-        })
+        # Final labels dicts for unidirectional search
+        self.finalLabel = None
 
         # If given, set REFs for dominance relations and feasibility checks
         if REF_forward and REF_backward:
@@ -250,6 +250,7 @@ class BiDirectional:
     # ALGORITHM #
     #############
     def _algorithm(self, direc):
+        log.debug("Direction = {}".format(direc))
         if direc == "forward":  # forward
             idx = 0  # index for head node
             # Update backwards half-way point
@@ -266,16 +267,10 @@ class BiDirectional:
         edges = deque(e for e in self.G.edges(data=True)
                       if e[idx] == self.currentLabel[direc].node)
         # If Label not been seen before, initialise a list
-        if (self.currentLabel[direc] not in self.unprocessedLabels[direc] and
-                self.currentLabel[direc] not in self.processedLabels[direc]):
-            self.unprocessedLabels[direc][self.currentLabel[direc]] = deque()
         # Propagate current label along all suitable edges in current direction
         list(map(self._propagate_label, edges, repeat(direc, len(edges))))
         # Extend label
         next_label = self._get_next_label(direc)
-        # If next label already been processed get another one.
-        if next_label in self.processedLabels[direc]:
-            next_label = self._get_next_label(direc, next_label)
         # Update current label
         self.currentLabel[direc] = next_label
         log.debug("current label = {}".format(next_label))
@@ -284,67 +279,29 @@ class BiDirectional:
         # Label propagation #
         # Get new label from current Label
         new_label = self.currentLabel[direc].get_new_label(edge, direc)
+        # If the new label is resource feasible
         if new_label and new_label.feasibility_check(self.max_res,
                                                      self.min_res):
-            # If label doesn't exist and is resource feasible
-            self.unprocessedLabels[direc][self.currentLabel[direc]].append(
-                new_label)
+            # And is not already in the unprocessed labels list
+            if (new_label not in self.unprocessedLabels[direc]
+                    and new_label not in self.processedLabels[direc]):
+                self.unprocessedLabels[direc].append(new_label)
 
     def _get_next_label(self, direc, exclude_label=None):
         # Label Extension #
-        try:
-            # Try to get labels create from the current one.
[…]
