### 2020-03-24 ded6af8
Updated half-way check and joining algorithm #25
