+++ cspy/algorithms/bidirectional.py
@@ -127,24 +127,24 @@ class BiDirectional:
 
         # Algorithm specific parameters #
         # Current forward and backward labels
-        self.currentLabel = OrderedDict({
+        self.current_label = OrderedDict({
             "forward":
             Label(0, "Source", zeros(G.graph["n_res"]), ["Source"]),
             "backward":
             Label(0, "Sink", max_res, ["Sink"])
         })
         # Unprocessed labels dict (both directions)
-        self.unprocessedLabels = OrderedDict({
+        self.unprocessed_labels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
         # Processed labels dict
-        self.processedLabels = OrderedDict({
+        self.processed_labels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
         # Final labels dicts for unidirectional search
-        self.finalLabel = None
+        self.final_label = None
 
         # If given, set REFs for dominance relations and feasibility checks
         if REF_forward and REF_backward:
@@ -167,7 +167,7 @@ class BiDirectional:
         """
         Calculate shortest path with resource constraints.
         """
-        while self.currentLabel["forward"] or self.currentLabel["backward"]:
+        while self.current_label["forward"] or self.current_label["backward"]:
             direc = self._get_direction()
             if direc:
                 self._algorithm(direc)
@@ -226,22 +226,22 @@ class BiDirectional:
     #############
     def _get_direction(self):
         if self.direc_in == "both":
-            if (self.currentLabel["forward"]
-                    and not self.currentLabel["backward"]):
+            if (self.current_label["forward"]
+                    and not self.current_label["backward"]):
                 return "forward"
-            elif (not self.currentLabel["forward"]
-                  and self.currentLabel["backward"]):
+            elif (not self.current_label["forward"]
+                  and self.current_label["backward"]):
                 return "backward"
-            elif (self.currentLabel["forward"]
-                  and self.currentLabel["backward"]):
+            elif (self.current_label["forward"]
+                  and self.current_label["backward"]):
                 return self.random_state.choice(["forward", "backward"])
             else:  # if both are empty
                 return
         else:
-            if (not self.currentLabel["forward"]
-                    and not self.currentLabel["backward"]):
+            if (not self.current_label["forward"]
+                    and not self.current_label["backward"]):
                 return
-            elif not self.currentLabel[self.direc_in]:
+            elif not self.current_label[self.direc_in]:
                 return
             else:
                 return self.direc_in
@@ -250,56 +250,55 @@ class BiDirectional:
     # ALGORITHM #
     #############
     def _algorithm(self, direc):
-        log.debug("Direction = {}".format(direc))
         if direc == "forward":  # forward
             idx = 0  # index for head node
             # Update backwards half-way point
             self.min_res[0] = max(
                 self.min_res[0],
-                min(self.currentLabel[direc].res[0], self.max_res[0]))
+                min(self.current_label[direc].res[0], self.max_res[0]))
         else:  # backward
             idx = 1  # index for tail node
             # Update forwards half-way point
             self.max_res[0] = min(
                 self.max_res[0],
-                max(self.currentLabel[direc].res[0], self.min_res[0]))
+                max(self.current_label[direc].res[0], self.min_res[0]))
         # Select edges with the same head/tail node as the current label node.
         edges = deque(e for e in self.G.edges(data=True)
-                      if e[idx] == self.currentLabel[direc].node)
+                      if e[idx] == self.current_label[direc].node)
         # If Label not been seen before, initialise a list
         # Propagate current label along all suitable edges in current direction
         list(map(self._propagate_label, edges, repeat(direc, len(edges))))
         # Extend label
         next_label = self._get_next_label(direc)
[…]
