### 2020-03-24 bac1e81
Patch25 (#26)
* Clean up bidirectional, joining alg #25

* Updated half-way check and joining algorithm #25

* Fixed broken link #15
