+++ cspy/algorithms/bidirectional.py
@@ -1,11 +1,12 @@
 from __future__ import absolute_import
 
+from copy import deepcopy
 from itertools import repeat
 from operator import add, sub
 from logging import getLogger
 from collections import OrderedDict, deque
 
-from numpy import zeros
+from numpy import zeros, array
 from numpy.random import RandomState
 
 # Local module imports
@@ -119,32 +120,31 @@ class BiDirectional:
                                   REF_backward)
         self.direc_in = direction
         self.max_res, self.min_res = max_res.copy(), min_res.copy()
-        self.max_res_in, self.min_res_in = max_res.copy(), min_res.copy()
+        self.max_res_in, self.min_res_in = array(max_res.copy()), array(
+            min_res.copy())
         # To expose results
         self.best_label = None
+
         # Algorithm specific parameters #
         # Current forward and backward labels
-        self.currentLabel = OrderedDict({
+        self.current_label = OrderedDict({
             "forward":
             Label(0, "Source", zeros(G.graph["n_res"]), ["Source"]),
             "backward":
             Label(0, "Sink", max_res, ["Sink"])
         })
         # Unprocessed labels dict (both directions)
-        self.unprocessedLabels = OrderedDict({
-            "forward": OrderedDict({}),
-            "backward": OrderedDict({})
-        })
-        # Processed labels dict
-        self.processedLabels = OrderedDict({
+        self.unprocessed_labels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
-        # Final labels dict.
-        self.finalLabel = OrderedDict({
-            "forward": self.currentLabel["forward"],
-            "backward": self.currentLabel["backward"]
+        # Processed labels dict
+        self.processed_labels = OrderedDict({
+            "forward": deque(),
+            "backward": deque()
         })
+        # Final labels dicts for unidirectional search
+        self.final_label = None
 
         # If given, set REFs for dominance relations and feasibility checks
         if REF_forward and REF_backward:
@@ -167,7 +167,7 @@ class BiDirectional:
         """
         Calculate shortest path with resource constraints.
         """
-        while self.currentLabel["forward"] or self.currentLabel["backward"]:
+        while self.current_label["forward"] or self.current_label["backward"]:
             direc = self._get_direction()
             if direc:
                 self._algorithm(direc)
@@ -226,22 +226,22 @@ class BiDirectional:
     #############
     def _get_direction(self):
         if self.direc_in == "both":
-            if (self.currentLabel["forward"]
-                    and not self.currentLabel["backward"]):
+            if (self.current_label["forward"]
+                    and not self.current_label["backward"]):
                 return "forward"
-            elif (not self.currentLabel["forward"]
-                  and self.currentLabel["backward"]):
+            elif (not self.current_label["forward"]
+                  and self.current_label["backward"]):
                 return "backward"
-            elif (self.currentLabel["forward"]
-                  and self.currentLabel["backward"]):
+            elif (self.current_label["forward"]
+                  and self.current_label["backward"]):
                 return self.random_state.choice(["forward", "backward"])
             else:  # if both are empty
                 return
         else:
-            if (not self.currentLabel["forward"]
-                    and not self.currentLabel["backward"]):
+            if (not self.current_label["forward"]
+                    and not self.current_label["backward"]):
                 return
-            elif not self.currentLabel[self.direc_in]:
+            elif not self.current_label[self.direc_in]:
                 return
             else:
                 return self.direc_in
@@ -255,96 +255,52 @@ class BiDirectional:
             # Update backwards half-way point
             self.min_res[0] = max(
                 self.min_res[0],
-                min(self.currentLabel[direc].res[0], self.max_res[0]))
[…]

[…]
