### 2020-03-24 72fc747
Codacy style fixes
