### 2020-03-24 torressa/cspy#15 @torressa
Added simple example that used to be in the paper as a jupyter notebook.
