### 2020-03-24 #1655 @torressa
Hi @stsievert !
Sorted.
I've added the simple example that used to be in the paper as a jupyter notebook.
I've explained the `n_res` parameter explicitly in the [Using cspy](https://cspy.readthedocs.io/en/latest/how_to.html) part of the docs and it I think it's pretty clear from the API examples what it is.
Hope that's ok
