### 2020-03-26 818d42c
Updated "join" algorithm in BiDirectional
 - Added stopping criteria for backward label comparisons.
 - Added filtering for forward labels using halway point.
 - Tested with maps instead of for-loops. Is a bit slower.
