### 2020-03-26 e5bca28
Updated changelog with new comparsions
