### 2020-04-01 2b424ea
Removed halway check in joining algorithm
