### 2020-04-03 #1655 @stsievert
Thanks for the improvements @torressa! The docs are looking better. Thanks for adding the simple example.

Could you add the simple example to the main documentation page? I think that'd help provide some more clarify.
