### 2020-04-03 e069815
Reduced search for joining algorithm
