### 2020-04-03 627d41b
Dev (#29)
* Removed halway check in joining algorithm

* Reduced search for joining algorithm

* Added vrpy submodule and custom ref fix

* Added jpath in docs and linked vrpy
