+++ cspy/algorithms/bidirectional.py
@@ -103,6 +103,7 @@ class BiDirectional:
     .. _Tilk 2017: https://www.sciencedirect.com/science/article/pii/S0377221717302035
     .. _Righini and Salani (2006): https://www.sciencedirect.com/science/article/pii/S1572528606000417
     """
+
     def __init__(self,
                  G,
                  max_res,
@@ -116,8 +117,7 @@ class BiDirectional:
         check(G, max_res, min_res, REF_forward, REF_backward, direction,
               __name__)
         # Preprocess graph
-        self.G = preprocess_graph(G, max_res, min_res, preprocess,
-                                  REF_backward)
+        self.G = preprocess_graph(G, max_res, min_res, preprocess, REF_backward)
         self.direc_in = direction
         self.max_res, self.min_res = max_res.copy(), min_res.copy()
         self.max_res_in, self.min_res_in = array(max_res.copy()), array(
@@ -128,18 +128,21 @@ class BiDirectional:
         # Algorithm specific parameters #
         # Current forward and backward labels
         self.current_label = OrderedDict({
-            "forward":
-            Label(0, "Source", zeros(G.graph["n_res"]), ["Source"]),
-            "backward":
-            Label(0, "Sink", max_res, ["Sink"])
+            "forward": Label(0, "Source", zeros(G.graph["n_res"]), ["Source"]),
+            "backward": Label(0, "Sink", max_res, ["Sink"])
         })
         # Unprocessed labels dict (both directions)
         self.unprocessed_labels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
-        # Processed labels dict
-        self.processed_labels = OrderedDict({
+        # All generated label
+        self.generated_labels = OrderedDict({
+            "forward": deque(),
+            "backward": deque()
+        })
+        # Non-dominated labels
+        self.nondominated_labels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
@@ -147,11 +150,13 @@ class BiDirectional:
         self.final_label = None
 
         # If given, set REFs for dominance relations and feasibility checks
-        if REF_forward and REF_backward:
+        if REF_forward:
             Label._REF_forward = REF_forward
-            Label._REF_backward = REF_backward
         else:
             Label._REF_forward = add
+        if REF_backward:
+            Label._REF_backward = REF_backward
+        else:
             Label._REF_backward = sub
         # Init with seed if given
         if seed is None:
@@ -226,20 +231,20 @@ class BiDirectional:
     #############
     def _get_direction(self):
         if self.direc_in == "both":
-            if (self.current_label["forward"]
-                    and not self.current_label["backward"]):
+            if (self.current_label["forward"] and
+                    not self.current_label["backward"]):
                 return "forward"
-            elif (not self.current_label["forward"]
-                  and self.current_label["backward"]):
+            elif (not self.current_label["forward"] and
+                  self.current_label["backward"]):
                 return "backward"
-            elif (self.current_label["forward"]
-                  and self.current_label["backward"]):
+            elif (self.current_label["forward"] and
+                  self.current_label["backward"]):
                 return self.random_state.choice(["forward", "backward"])
             else:  # if both are empty
                 return
         else:
-            if (not self.current_label["forward"]
-                    and not self.current_label["backward"]):
+            if (not self.current_label["forward"] and
+                    not self.current_label["backward"]):
                 return
             elif not self.current_label[self.direc_in]:
                 return
@@ -280,8 +285,8 @@ class BiDirectional:
         if new_label and new_label.feasibility_check(self.max_res,
                                                      self.min_res):
             # And is not already in the unprocessed labels list
-            if (new_label not in self.unprocessed_labels[direc]
-                    and new_label not in self.processed_labels[direc]):
+            if (new_label not in self.unprocessed_labels[direc] and
[…]
