+++ docs/how_to.rst
@@ -60,7 +60,14 @@ For this reason, it sometimes takes longer than the others.
 The remaining algorithms are metaheuristics,
 i.e. they provide fast and approximate solutions to the CSP problem.
 
-Please see the `examples`_ for more in-depth usage.
+Examples
+~~~~~~~~
+
+The following examples are included in the `examples`_ for more in-depth usage.
+
+- `jpath`_ : Simple example showing the necessary graph adptations and the use of custom resource extension functions. Also discussed below.
+- `vrpy`_: (under development) external vehicle routing framework which uses ``cspy`` to solve different variants of the vehicle routing problem using column generation.
+- `cgar`_: Complex example using ``cspy`` for column generation applied to the aircraft recovery problem.
 
 Please see individual algorithm documentation for simple examples.
 
@@ -71,9 +78,13 @@ Please see individual algorithm documentation for simple examples.
 .. _GRASP: https://cspy.readthedocs.io/en/latest/api/cspy.GRASP.html
 .. _Marinakis et al 2017: https://www.sciencedirect.com/science/article/pii/S0377221717302357z
 .. _examples: https://github.com/torressa/cspy/tree/master/examples/
+.. _vrpy: https://github.com/Kuifje02/vrpy
+
+REFs
+~~~~
 
 Pre-requirements
-~~~~~~~~~~~~~~~~
+****************
 
 For the :class:`BiDirectional` algorithm, there is a number of assumptions required by definition (`Tilk et al 2017`_).
 
@@ -102,9 +113,6 @@ For assumption 2, if resource extension functions are additive, these are easily
 However, when using custom resource extension functions (discussed below),
 it is up to the user to define them appropriately!
 
-REFs
-~~~~
-
 Additive REFs
 *************
 
@@ -129,8 +137,7 @@ However, it is up the users to ensure that the custom REFs are well defined,
 it may be the case that the algorithm fails to find a feasible path, or gets stuck.
 
 For theoretical information on what REFs are we refer you to the paper by `Inrich 2005`_.
-For a brief overview with a practical implementation see the `examples`_ and 
-the pdf document `cgar`_.
+For a brief overview with a practical implementation see any of the `examples`_.
 
 Custom REF template
 *******************
@@ -155,7 +162,123 @@ As an example, suppose the 2nd resource represents travel time (``res[1]``). Sup
 
 Your custom REF can then be passed with this format, into the algorithm of choice using the ``REF`` argument (see individual algorithms for details). Note that for the :class:`BiDirectional` algorithm, due to the properties of the algorithm, if you want to use this feature, you have to pass two custom REFs: one for the forward search and one for the backward search. Where the backward REF has to be the inverse of the forward REF, otherwise the algorithm will not return a meaningful path (`Tilk et al 2017`_). It is up to the user to ensure this is the case.
 
-.. _cgar: https://github.com/torressa/cspy/examples/cgar/cgar.pdf
+
+Simple Example
+~~~~~~~~~~~~~~
+
+For illustration of most of the things discussed above, consider the following example.
+
+Jane is part-time postwoman working in Delft, Netherlands. However, she is assigned a small area (the Indische Buurt neighbourhood) so when planning her daily route she wants to make it as long and exciting as possible. 
+That is, when planning her routes she has to consider the total shift time, sights visited, travel time, and delivery time. Her shift has to be at most 5 hours.
+
+
+This problem can easily be modelled as a CSP problem. 
+With the description above, the set of resources can be defined as,
+
+```python
+R = ['sights', 'shift', 'travel-time', 'delivery-time'] 
+# len(R) = 4
+```
[…]
