+++ docs/how_to.rst
@@ -176,8 +176,9 @@ This problem can easily be modelled as a CSP problem.
 With the description above, the set of resources can be defined as,
 
 .. code-block:: python
-    R = ['sights', 'shift', 'travel-time', 'delivery-time'] 
-    # len(R) = 4
+
+        R = ['sights', 'shift', 'travel-time', 'delivery-time'] 
+        # len(R) = 4
 
 Let ``G`` denote a directed graph with edges to/from all streets of the Indische Buurt 
 neighbourhood. 
@@ -188,18 +189,20 @@ The first entry of this array, corresponds to the ``'sights'`` resource, i.e. ho
 Also, when defining ``G``, one has to specify the number of resources ``n_res``, which also has to be equal to ``len(R)``.
 
 .. code-block:: python
-    from networkx import DiGraph
-    G = DiGraph(directed=True, n_res=4)  # init network
+
+        from networkx import DiGraph
+        G = DiGraph(directed=True, n_res=4)  # init network
 
 Now, using the open source package OSMnx, we can easily generate a network for Jane's neighbourhood
 
 .. code-block:: python
-    from osmnx import graph_from_address, plot_graph
 
-    M = graph_from_address('Ceramstraat, Delft, Netherlands',
-                               distance=1600,
-                               network_type='walk',
-                               simplify=False)
+        from osmnx import graph_from_address, plot_graph
+
+        M = graph_from_address('Ceramstraat, Delft, Netherlands',
+                                   distance=1600,
+                                   network_type='walk',
+                                   simplify=False)
 
 We have to transform the network for one compatible with cspy.
 To do this suppose we have two functions from ``jpath_preprocessing`` 
@@ -207,18 +210,19 @@ that perform all the changes required
 (for more details, see `jpath`_)
 
 .. code-block:: python
-    from networkx import DiGraph
-    from jpath_preprocessing import relabel_source_sink, add_cspy_edge_attributes
 
-    # Transform M to comply with cspy's prerequirements
-    # Convert MultiGraph into a Digraph with attribute 'n_res'
-    G = DiGraph(M, directed=True, n_res=5)
-    # Relabel source node to "Source" and sink node to "Sink" (see function for more details)
-    G = relabel_source_sink(G)
-    # Add res_cost and other resource attributes (see function for more details)
-    G = add_cspy_edge_attributes(G)
+        from networkx import DiGraph
+        from jpath_preprocessing import relabel_source_sink, add_cspy_edge_attributes
 
-    n_edges = len(G.edges())  # number of edges in network
+        # Transform M to comply with cspy's prerequirements
+        # Convert MultiGraph into a Digraph with attribute 'n_res'
+        G = DiGraph(M, directed=True, n_res=5)
+        # Relabel source node to "Source" and sink node to "Sink" (see function for more details)
+        G = relabel_source_sink(G)
+        # Add res_cost and other resource attributes (see function for more details)
+        G = add_cspy_edge_attributes(G)
+
+        n_edges = len(G.edges())  # number of edges in network
 
 To define the custom REFs,  ``jane_REF``, that controls how resources evolve throughout the path,
 we require two inputs: an array of current cumulative resource values ``res``, 
@@ -226,20 +230,21 @@ and the edge that is being considered for an extension of a path ``edge``
 (which consists of two nodes and the edge data).
 
 .. code-block:: python
-    from numpy import array
-    def jane_REF(res, edge):
-        arr = array(res)  # local array
[…]
