### 2020-04-03 #1655 @torressa
Done! Thanks @stsievert for the suggestions.
By the way, more examples coming!
I'm collaborating with someone to implements some vehicle routing variants using cspy!!
