### 2020-04-03 d379e4a
Update dev (#31)
* Dev (#29)

* Removed halway check in joining algorithm

* Reduced search for joining algorithm

* Added vrpy submodule and custom ref fix

* Added jpath in docs and linked vrpy

* Added code-blocks in docs

* Removed conflicts

* Docs code-block formatting
