+++ docs/how_to.rst
@@ -175,10 +175,11 @@ That is, when planning her routes she has to consider the total shift time, sigh
 This problem can easily be modelled as a CSP problem. 
 With the description above, the set of resources can be defined as,
 
-```python
-R = ['sights', 'shift', 'travel-time', 'delivery-time'] 
-# len(R) = 4
-```
+.. code-block:: python
+
+        R = ['sights', 'shift', 'travel-time', 'delivery-time'] 
+        # len(R) = 4
+
 
 Let ``G`` denote a directed graph with edges to/from all streets of the Indische Buurt 
 neighbourhood. 
@@ -188,63 +189,68 @@ The entries of ``res_cost`` have the same order as the entries in ``R``.
 The first entry of this array, corresponds to the ``'sights'`` resource, i.e. how many sights there are along a specific edge. The last entry of this array, corresponds to the ``'delivery-time'`` resource, i.e. time taken to deliver post along a specific edge. The remaining entries can be initialised to be 0.
 Also, when defining ``G``, one has to specify the number of resources ``n_res``, which also has to be equal to ``len(R)``.
 
-```python
-from networkx import DiGraph
-G = DiGraph(directed=True, n_res=4)  # init network
-```
+
+.. code-block:: python
+
+        from networkx import DiGraph
+        G = DiGraph(directed=True, n_res=4)  # init network
 
 Now, using the open source package OSMnx, we can easily generate a network for Jane's neighbourhood
 
-```python
-from osmnx import graph_from_address, plot_graph
+.. code-block:: python
 
-M = graph_from_address('Ceramstraat, Delft, Netherlands',
-                           distance=1600,
-                           network_type='walk',
-                           simplify=False)
-```
+        from osmnx import graph_from_address, plot_graph
+
+        M = graph_from_address('Ceramstraat, Delft, Netherlands',
+                                   distance=1600,
+                                   network_type='walk',
+                                   simplify=False)
 
 We have to transform the network for one compatible with cspy.
 To do this suppose we have two functions from ``jpath_preprocessing`` 
 that perform all the changes required 
 (for more details, see `jpath`_)
 
-```python
-from networkx import DiGraph
-from jpath_preprocessing import relabel_source_sink, add_cspy_edge_attributes
 
-# Transform M to comply with cspy's prerequirements
-# Convert MultiGraph into a Digraph with attribute 'n_res'
-G = DiGraph(M, directed=True, n_res=5)
-# Relabel source node to "Source" and sink node to "Sink" (see function for more details)
-G = relabel_source_sink(G)
-# Add res_cost and other resource attributes (see function for more details)
-G = add_cspy_edge_attributes(G)
+.. code-block:: python
 
-n_edges = len(G.edges())  # number of edges in network
-```
+        from networkx import DiGraph
+        from jpath_preprocessing import relabel_source_sink, add_cspy_edge_attributes
+
+        # Transform M to comply with cspy's prerequirements
+        # Convert MultiGraph into a Digraph with attribute 'n_res'
+        G = DiGraph(M, directed=True, n_res=5)
+        # Relabel source node to "Source" and sink node to "Sink" (see function for more details)
+        G = relabel_source_sink(G)
+        # Add res_cost and other resource attributes (see function for more details)
+        G = add_cspy_edge_attributes(G)
+
+        n_edges = len(G.edges())  # number of edges in network
+
 
[…]
