### 2020-04-06 eeff68a
Added filtering for join, generalised dominance
