+++ cspy/algorithms/bidirectional.py
@@ -141,8 +141,8 @@ class BiDirectional:
             "forward": deque(),
             "backward": deque()
         })
-        # Non-dominated labels
-        self.nondominated_labels = OrderedDict({
+        # To save all best labels
+        self.best_labels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
@@ -176,7 +176,6 @@ class BiDirectional:
             direc = self._get_direction()
             if direc:
                 self._algorithm(direc)
-                self._check_dominance(direc)
             else:
                 break
         return self._process_paths()
@@ -239,6 +238,7 @@ class BiDirectional:
                 return "backward"
             elif (self.current_label["forward"] and
                   self.current_label["backward"]):
+                # TODO implement choices based on number of labels
                 return self.random_state.choice(["forward", "backward"])
             else:  # if both are empty
                 return
@@ -277,6 +277,8 @@ class BiDirectional:
         next_label = self._get_next_label(direc)
         # Update current label
         self.current_label[direc] = next_label
+        # Dominance checks
+        self._check_dominance(next_label, direc)
 
     def _propagate_label(self, edge, direc):
         # Label propagation #
@@ -295,7 +297,7 @@ class BiDirectional:
         current_label = self.current_label[direc]
         unproc_labels = self.unprocessed_labels[direc]
         self.generated_labels[direc].append(current_label)
-        self._remove_unprocessed_labels([current_label], direc)
+        self._remove_labels([current_label], direc, unproc=True)
         # Return label with minimum monotone resource for the forward search
         # and the maximum monotone resource for the backward search
         if unproc_labels:
@@ -309,42 +311,55 @@ class BiDirectional:
     #############
     # DOMINANCE #
     #############
-    def _check_dominance(self, direc):
+    def _check_dominance(self, label_to_check, direc, unproc=True, best=False):
         """
-        For all labels, check if it is dominated, or itself dominates other
-        labels. If this is found to be the case, the dominated label is
-        removed.
+        For all labels, checks if ``label_to_check`` is dominated,
+        or itself dominates any other label in either the unprocessed_labels
+        list or the non-dominated labels list.
+        If this is found to be the case, the dominated label(s) is(are)
+        removed from the appropriate list.
         """
-        # Reference attributes
-        current_label = self.current_label[direc]
-        unproc_labels = self.unprocessed_labels[direc]
-        # If current label is not None (at termination)
-        if current_label:
-            keys_to_pop = deque()
+        # Select appropriate list to check
+        if unproc:
+            labels_to_check = self.unprocessed_labels[direc]
+        elif best:
+            labels_to_check = self.best_labels[direc]
+        # If label is not None (at termination)
+        if label_to_check:
+            labels_to_pop = deque()
             # Gather all comparable labels (same node)
             all_labels = deque(
-                lab for lab in unproc_labels
-                if lab.node == current_label.node and lab != current_label)
+                l for l in labels_to_check
+                if l.node == label_to_check.node and l != label_to_check)
             # Add to list for removal if they are dominated
-            keys_to_pop.extend(lab for lab in all_labels
-                               if current_label.dominates(lab, direc))
-            # Add current label for removal if itself is dominated
-            if any(lab.dominates(current_label, direc) for lab in all_labels):
-                keys_to_pop.append(current_label)
-            # If not bidirectional algorithm, check and save current label
-            else:
+            labels_to_pop.extend(
+                l for l in all_labels if label_to_check.dominates(l, direc))
+            # Add input label for removal if itself is dominated
[…]
