### 2020-04-06 ffcae93
Added custom REF inversion
