### 2020-04-07 67df376
Updated docstring, removed time-out test #32
