### 2020-04-08 713005a
Removed negative cost cycle check for bidirectional
