### 2020-04-08 7f49b5a
Merge master into dev (#34)
* Dev (#29)

* Removed halway check in joining algorithm

* Reduced search for joining algorithm

* Added vrpy submodule and custom ref fix

* Added jpath in docs and linked vrpy

* Added code-blocks in docs

* Removed conflicts

* Docs code-block formatting

* Added filtering for join, generalised dominance

* Added custom REF inversion

* Removed backward REF for Bidirectional

* Updated docstring, removed time-out test #32
