+++ cspy/algorithms/bidirectional.py
@@ -2,11 +2,11 @@ from __future__ import absolute_import
 
 from copy import deepcopy
 from itertools import repeat
-from operator import add, sub
+from operator import add
 from logging import getLogger
 from collections import OrderedDict, deque
 
-from numpy import zeros, array
+from numpy import array
 from numpy.random import RandomState
 
 # Local module imports
@@ -26,7 +26,6 @@ class BiDirectional:
     Depending on the range of values for U, L, we get
     four different algorithms. See self.name_algorithm and Notes.
 
-
     Parameters
     ----------
     G : object instance :class:`nx.Digraph()`
@@ -43,9 +42,9 @@ class BiDirectional:
         usage (including initial backward stopping point).
         We must have ``len(min_res)`` :math:`=` ``len(max_res)`` :math:`\geq 2`
 
-    REF_forward REF_backward : function, optional
+    REF : function, optional
         Custom resource extension function. See `REFs`_ for more details.
-        Default : additive, subtractive.
+        Default : additive.
 
     preprocess : bool, optional
         enables preprocessing routine. Default : False.
@@ -108,16 +107,14 @@ class BiDirectional:
                  G,
                  max_res,
                  min_res,
-                 REF_forward=None,
-                 REF_backward=None,
+                 REF=None,
                  preprocess=False,
                  direction="both",
                  seed=None):
         # Check inputs and preprocess G unless option disabled
-        check(G, max_res, min_res, REF_forward, REF_backward, direction,
-              __name__)
+        check(G, max_res, min_res, REF, direction, __name__)
         # Preprocess graph
-        self.G = preprocess_graph(G, max_res, min_res, preprocess, REF_backward)
+        self.G = preprocess_graph(G, max_res, min_res, preprocess, REF)
         self.direc_in = direction
         self.max_res, self.min_res = max_res.copy(), min_res.copy()
         self.max_res_in, self.min_res_in = array(max_res.copy()), array(
@@ -126,10 +123,13 @@ class BiDirectional:
         self.best_label = None
 
         # Algorithm specific parameters #
+        # set bounds for bacward search
+        bwd_start = deepcopy(min_res)
+        bwd_start[0] = max_res[0]
         # Current forward and backward labels
         self.current_label = OrderedDict({
-            "forward": Label(0, "Source", zeros(G.graph["n_res"]), ["Source"]),
-            "backward": Label(0, "Sink", max_res, ["Sink"])
+            "forward": Label(0, "Source", min_res, ["Source"]),
+            "backward": Label(0, "Sink", bwd_start, ["Sink"])
         })
         # Unprocessed labels dict (both directions)
         self.unprocessed_labels = OrderedDict({
@@ -141,8 +141,8 @@ class BiDirectional:
             "forward": deque(),
             "backward": deque()
         })
-        # Non-dominated labels
-        self.nondominated_labels = OrderedDict({
+        # To save all best labels
+        self.best_labels = OrderedDict({
             "forward": deque(),
             "backward": deque()
         })
@@ -150,14 +150,10 @@ class BiDirectional:
         self.final_label = None
 
         # If given, set REFs for dominance relations and feasibility checks
-        if REF_forward:
-            Label._REF_forward = REF_forward
+        if REF:
+            Label._REF = REF
         else:
-            Label._REF_forward = add
-        if REF_backward:
-            Label._REF_backward = REF_backward
-        else:
-            Label._REF_backward = sub
+            Label._REF = add
         # Init with seed if given
         if seed is None:
             self.random_state = RandomState()
@@ -176,7 +172,6 @@ class BiDirectional:
             direc = self._get_direction()
             if direc:
                 self._algorithm(direc)
-                self._check_dominance(direc)
             else:
[…]

[…]
