### 2020-04-08 44eac83
Removed bididerection exception raise from check
