### 2020-04-08 1eefa1e
Merge dev into master (#35)
* Update dev (#31)

* Dev (#29)

* Removed halway check in joining algorithm

* Reduced search for joining algorithm

* Added vrpy submodule and custom ref fix

* Added jpath in docs and linked vrpy

* Added code-blocks in docs

* Removed conflicts

* Docs code-block formatting

* Removed negative cost cycle check for bidirectional

* Merge master into dev (#34)

* Dev (#29)

* Removed halway check in joining algorithm

* Reduced search for joining algorithm

* Added vrpy submodule and custom ref fix

* Added jpath in docs and linked vrpy

* Added code-blocks in doc
[…]
