+++ cspy/checking.py
@@ -133,7 +133,7 @@ def _check_path(G, max_res, min_res, direction, algorithm):
     try:
         if not has_path(G, 'Source', 'Sink'):
             raise NetworkXException("Disconnected Graph")
-        if negative_edge_cycle(G):
+        if negative_edge_cycle(G) and "bidirectional" not in algorithm:
             raise NetworkXException("A negative cost cycle was found")
     except NetworkXException as e:
         raise Exception("An error occurred: {}".format(e))
