### 2020-04-09 c3318b7
Merge branch 'dev' into master (#36)
* Update dev (#31)

* Dev (#29)

* Removed halway check in joining algorithm

* Reduced search for joining algorithm

* Added vrpy submodule and custom ref fix

* Added jpath in docs and linked vrpy

* Added code-blocks in docs

* Removed conflicts

* Docs code-block formatting

* Removed negative cost cycle check for bidirectional

* Merge master into dev (#34)

* Dev (#29)

* Removed halway check in joining algorithm

* Reduced search for joining algorithm

* Added vrpy submodule and custom ref fix

* Added jpath in docs and linked vrpy

* Added code-bloc
[…]
