### 2020-04-09 a092611
Removed functions in label, simplified bidirect
