+++ CHANGELOG.md
@@ -7,13 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+
+- BiDirectional:
+  - Option to chose method for direction selection.
+- [vrpy](https://github.com/Kuifje02/vrpy) submodule.
+
 ### Changed 
 
-- BiDirectional
+- BiDirectional:
   - Label storage, divided into unprocessed, generated and non-dominated labels
-  - Restricted join algorithm to non-dominated labels 
+  - Restricted join algorithm to non-dominated label
   - Changed backward resource extensions to avoid complex and computationally costly inversion. Additionally, it removes the requirement of an explicit backward REF.
-
+  - Filtering for backward labels in join algorithm.
+  - Cleaned up unused label operator overloads.
+  - Removed costly comparison in `_propagate_label`.
+  - Changed generated labels attributes from dict of deques to dict of int with count.
 
 ## [v0.0.14]
 
@@ -143,8 +152,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docstring modifications to include maths.
 - Updated README.
 
-[unreleased]: https://github.com/torressa/cspy/compare/v0.0.14...HEAD
-[v0.0.14]: https://github.com/torressa/cspy/compare/v0.0.13...v0.0.14
+[unreleased]: https://github.com/torressa/cspy/compare/v0.0.13...HEAD
 [v0.0.13]: https://github.com/torressa/cspy/compare/v0.0.12...v0.0.13
 [v0.0.12]: https://github.com/torressa/cspy/compare/v0.0.11...v0.0.12
 [v0.0.11]: https://github.com/torressa/cspy/compare/v0.0.10...v0.0.11
