### 2020-04-14 cc31c8f
Removed negative edge cycle assumption
- Rework of path and algorithm attributes to avoid duplication
- Replaced astar algorithm with a procedure that finds a short simple
path (tabu and greedy_elimination)

Updated added stopping criteria for tabu
