+++ tests/tests_issue20.py
@@ -3,6 +3,7 @@ import unittest
 
 from numpy import array
 from networkx import DiGraph
+from parameterized import parameterized
 
 sys.path.append("../")
 from cspy.algorithms.tabu import Tabu
@@ -11,9 +12,10 @@ from cspy.algorithms.bidirectional import BiDirectional
 
 class TestsIssue20(unittest.TestCase):
     """
-    Tests for issue #20 
+    Tests for issue #20
     https://github.com/torressa/cspy/issues/20
     """
+
     def setUp(self):
         # Create simple digraph with appropriate attributes
         self.G = DiGraph(directed=True, n_res=2)
@@ -28,11 +30,12 @@ class TestsIssue20(unittest.TestCase):
 
         self.max_res, self.min_res = [len(self.G.edges()), 2], [0, 0]
 
-    def testBiDirectional(self):
+    @parameterized.expand(zip(range(100), range(100)))
+    def testBiDirectional(self, _, seed):
         """
         Find shortest path of simple test digraph using BiDirectional
         """
-        bidirec = BiDirectional(self.G, self.max_res, self.min_res)
+        bidirec = BiDirectional(self.G, self.max_res, self.min_res, seed=seed)
         bidirec.run()
         path = bidirec.path
         cost = bidirec.total_cost
@@ -54,10 +57,10 @@ class TestsIssue20(unittest.TestCase):
         cost = tabu.total_cost
         total_res = tabu.consumed_resources
         # Check attributes
-        self.assertEqual(cost, 0)
-        self.assertTrue(all(total_res == [2, 1]))
+        self.assertEqual(cost, -5)
+        self.assertTrue(all(total_res == [3, 2]))
+        self.assertEqual(path, ['Source', 3, 2, 'Sink'])
         # Check path
-        self.assertEqual(path, ['Source', 1, 'Sink'])
         self.assertTrue(all(e in self.G.edges() for e in zip(path, path[1:])))
 
 
