### 2020-04-18 c5cf26d
Added algorithm option in tabu and greedelim
