+++ cspy/algorithms/tabu.py
@@ -38,13 +38,25 @@ class Tabu(PathBase):
     preprocess : bool, optional
         enables preprocessing routine. Default : False.
 
+    algorithm : string, optional
+        shortest path algorithm to use. Options are "`simple`_"
+        or "`astar`_".
+        If the input network has a negative cycle, the "`simple`_"
+        algorithm is automatically chosen (as the astar algorithm cannot cope).
+
     max_depth : int, optional
         depth for search of shortest simple path. Default : 1000.
         If the total number of simple paths is less than max_depth,
         then the shortest path is used.
 
     .. _REFs : https://cspy.readthedocs.io/en/latest/how_to.html#refs
+    .. _simple : https://networkx.github.io/documentation/stable/reference/algorithms/generated/networkx.algorithms.simple_paths.shortest_simple_paths.html#networkx.algorithms.simple_paths.shortest_simple_paths
+    .. _astar : https://networkx.github.io/documentation/stable/reference/algorithms/generated/networkx.algorithms.shortest_paths.astar.astar_path.html#networkx.algorithms.shortest_paths.astar.astar_path
 
+    Raises
+    ------
+    Exception
+        if no resource feasible path is found
 
     Notes
     -----
@@ -89,9 +101,10 @@ class Tabu(PathBase):
                  min_res,
                  REF=None,
                  preprocess=False,
+                 algorithm="simple",
                  max_depth=1000):
         # Pass arguments to SimplePath object
-        super().__init__(G, max_res, min_res, REF, preprocess)
+        super().__init__(G, max_res, min_res, REF, preprocess, algorithm)
         # Algorithm specific parameters
         self.max_depth = max_depth
         self.iteration = 0
@@ -117,7 +130,7 @@ class Tabu(PathBase):
     def _algorithm(self):
         path = []
         try:
-            path = self.update_simple_path(self.neighbour, self.max_depth)
+            path = self.get_shortest_path(self.neighbour, self.max_depth)
         except NetworkXException:
             pass
         if path:
