### 2020-04-18 b700593
Changed halway-check (to actually do something!)
