### 2020-04-25 5e8c0ae
Fixed #38, moved notes+examples to docs
