### 2020-04-25 6851494
Dev (#39)
* Resurrected backward REF, added REF join argument

* Fixed #38, moved notes+examples to docs

* Removed log
