+++ cspy/algorithms/bidirectional.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 from copy import deepcopy
 from itertools import repeat
-from operator import add
+from operator import add, sub
 from logging import getLogger
 from collections import OrderedDict, deque
 
@@ -21,8 +21,8 @@ class BiDirectional:
     """
     Implementation of the bidirectional labeling algorithm with dynamic
     half-way point (`Tilk 2017`_).
-    This requires the half-way procedure from `Righini and Salani (2006)`_,
-    also implemented.
+    This requires the joining procedure (Algorithm 3) from
+    `Righini and Salani (2006)`_, also implemented.
     Depending on the range of values for U, L, we get
     four different algorithms. See self.name_algorithm and Notes.
 
@@ -42,10 +42,6 @@ class BiDirectional:
         usage (including initial backward stopping point).
         We must have ``len(min_res)`` :math:`=` ``len(max_res)`` :math:`\geq 2`
 
-    REF : function, optional
-        Custom resource extension function. See `REFs`_ for more details.
-        Default : additive.
-
     preprocess : bool, optional
         enables preprocessing routine. Default : False.
 
@@ -64,48 +60,11 @@ class BiDirectional:
         seed for PSOLGENT class. Default : None (which gives a single value
         numpy.random.RandomState).
 
+    REF_forward, REF_backward, REF_join : functions, optional
+        Custom resource extension functions. See `REFs`_ for more details.
+        Default : additive.
+
     .. _REFs : https://cspy.readthedocs.io/en/latest/how_to.html#refs
-
-    Notes
-    -----
-    The input graph must have a ``n_res`` attribute which must be
-    :math:`\geq 2`. The edges in the graph must all have a ``res_cost``
-    attribute.
-
-    According to the inputs, four different algorithms can be used.
-    If you'd like to check which algorithm is running given the resource
-    limits, call the method :func:`BiDirectional.name_algorithm()`
-    for a log with the classification.
-
-    - ``direction`` = "forward": Monodirectional forward labeling algorithm
-    - :math:`H_F == H_B`: Bidirectional labeling algorithm with static halfway point.
-    - ``direction`` = "backward": Monodirectional backward labeling algorithm
-    - :math:`H_F > H_B`: Bidirectional labeling algorithm with dynamic halfway point.
-    - :math:`H_F < H_B`: The algorithm won't go anywhere!
-
-    Example
-    -------
-    To run the algorithm, create a :class:`BiDirectional` instance and call
-    ``run``.
-
-    .. code-block:: python
-
-        >>> from cspy import BiDirectional
-        >>> from networkx import DiGraph
-        >>> from numpy import array
-        >>> G = DiGraph(directed=True, n_res=2)
-        >>> G.add_edge("Source", "A", res_cost=array([1, 2]), weight=0)
-        >>> G.add_edge("A", "B", res_cost=array([1, 0.3]), weight=0)
-        >>> G.add_edge("A", "C", res_cost=array([1, 0.1]), weight=0)
-        >>> G.add_edge("B", "C", res_cost=array([1, 3]), weight=-10)
-        >>> G.add_edge("B", "Sink", res_cost=array([1, 2]), weight=10)
-        >>> G.add_edge("C", "Sink", res_cost=array([1, 10]), weight=0)
-        >>> max_res, min_res = [4, 20], [1, 0]
-        >>> bidirec = BiDirectional(G, max_res, min_res, direction="both")
-        >>> bidirec.run()
-        >>> print(bidirec.path)
-        ["Source", "A", "B", "C", "Sink"]
-
     .. _Tilk 2017: https://www.sciencedirect.com/science/article/pii/S0377221717302035
[…]

[…]
