### 2020-04-25 c5a5323
Updated changelog, bumped version docs conf file
