### 2020-04-25 2285ed9
Removed unused import and useless-else-on-loop
