### 2020-04-26 93d1dd0
Updated examples in the paper
