+++ paper/paper.md
@@ -59,12 +59,17 @@ A key component of a CSP problem is to define a set of resources and a function 
 
 # Examples
 
+The package has been used in the following examples:
+ - [`vrpy`](https://github.com/Kuifje02/vrpy) : vehicle routing framework which solves different variants of the vehicle routing problem (including capacity constraints and time-windows) using column generation. The framework has been tested on standard vehicle routing instances.
+ - [`cgar`](https://github.com/torressa/cspy/tree/master/examples/cgar) : Complex example using column generation applied to the aircraft recovery problem.
+ -[`jpath`](https://github.com/torressa/cspy/tree/master/examples/jpath) : Simple example showing the necessary graph adaptations and the use of custom resource extension functions.
 For some real-world examples, see the repository.
 
-Currently included is an implementation of column generation for the aircraft recovery problem. 
 
 # Acknowledgements
 
 The author gratefully acknowledges the support of the EPSRC funded EP/L015692/1 STOR-i Centre for Doctoral Training. 
+
+I would also like to thank Romain Montagné for his initiative and time invested to write [`vrpy`](https://github.com/Kuifje02/vrpy).
 
 # References
