### 2020-04-26 0ce4936
Paper updated. Removed spacing in example bullets
