### 2020-04-26 6e85ac5
Paper updated. Added spacing for example bullets
