### 2020-04-26 #1655 @torressa
Just updated the paper with a list of linked examples.
Sorry for the bombarding, couldn't get the bullets to work
