### 2020-04-29 torressa/cspy#15 @stsievert
Thanks for the simple example at https://cspy.readthedocs.io/en/latest/how_to.html#simple-example. That clears up a lot of the usage.

> We have to transform the network for one compatible with cspy. To do this suppose we have two functions from jpath_preprocessing that perform all the changes required (for more details, see jpath)

This seems important enough that a description should be included in the example. It could be short, maybe only a sentence. I'd like for the user to better understand the requirements.

Maybe something like this?

> We have to transform the network for one compatible with cspy, as per the [Input requirements][1]. The following code will convert a city map into a directed graph, rename the start/end nodes of Janes walk to be `Source` and `Sink` (names which cspy uses), and calculate the specifics of Jane's walk (figuring out travel time, adding buildings/sights, etc).
>
>  ``` python
>
> from networkx import DiGraph
>
> # Transform M to comply...
> G = DiGraph(M, directed=True, n_res=5)
>
> # Relabel nodes the start/end nodes as source/sink
>  G = jpath_preprocessing.relabel_source_sink(G, {"Source": "Ternatestraat", "Sink": "Delftweg"})
>
> # Figure out details of Jane's walk
>  G = jpath_preprocessing.add_cspy_edge_attributes(G)
> ```

[1]:https://cspy.readthedocs.io/en/latest/how_to.html#input-requirements
