### 2020-04-29 #1655 @stsievert
I've checked off all my boxes. The simple example at https://cspy.readthedocs.io/en/latest/how_to.html#simple-example has greatly helped improve the usability of this software.

I have unchecked the software versioning box. I would have not accepted v0.0.8; however I would be willing to accept the current release v0.1.0 (perhaps in part due to https://github.com/torressa/cspy/issues/14, https://github.com/torressa/cspy/issues/15 and https://github.com/torressa/cspy/issues/16). How should that be managed?
