### 2020-04-29 6ad0650
Add more clear explanation in docs #15
