### 2020-04-29 #1655 @torressa
Just updated the docs.

Versioning-wise, do you mean that the points need updating to match the current version (v0.1.0)?

Since the start of the submission there has a fair amount of bug fixes (around 6 major ones), a few new features and some other changes stemming from the development of the vehicle routing application [vrpy](https://github.com/Kuifje02/vrpy/). Taking the version to v0.0.14.

After the exact algorithm is actually providing known optimal solutions to the vehicle routing problem I decided to bump from v0.0.14 to v0.1.0.
