### 2020-04-29 #1655 @stsievert
>> I would have not accepted v0.0.8; however I would be willing to accept the current release v0.1.0 (perhaps in part due to torressa/cspy#14, torressa/cspy#15 and torressa/cspy#16). How should that be managed?
>
> Versioning-wise, do you mean that the points need updating to match the current version (v0.1.0)? Since the start of the submission there has a fair amount of bug fixes

Whoops, thanks. I meant to ask @poulson or @afron that question.
