### 2020-05-01 3a53383
Remove deque map and swo file
