### 2020-05-07 509fbf5
Add test for issue and fix for some cases of #41
