### 2020-05-07 c4f3ba6
Uncomment mono directional tests
