### 2020-05-21 0bc7850
Update CHANGELOG for JOSS release
