### 2020-05-21 c681681
Merger 'master' into 'dev' (#44)
* Dev (#39)

* Resurrected backward REF, added REF join argument

* Fixed #38, moved notes+examples to docs

* Removed log

* Updated changelog, bumped version docs conf file

* squash! Dev (#39)

* Removed unused import and useless-else-on-loop

* Updated examples in the paper

* Paper updated. Removed spacing in example bullets

* Paper updated. Added spacing for example bullets

* Add more clear explanation in docs #15

* Remove deque map and swo file

* Update commit vrpy submodule

* Update CHANGELOG for JOSS release
