+++ docs/conf.py
@@ -14,15 +14,13 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('../'))
 
-
 # -- Project information -----------------------------------------------------
 project = 'cspy'
 copyright = '2019, David Torres Sanchez'
 author = 'David Torres Sanchez'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.1'
-
+release = '0.1.0'
 
 # -- General configuration ---------------------------------------------------
 
@@ -38,7 +36,6 @@ templates_path = ['_templates']
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
-
 
 # -- Options for HTML output -------------------------------------------------
 
